% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adressed_demand.R
\name{adressed_demand}
\alias{adressed_demand}
\title{Calculer la demande adressée à chaque pays}
\usage{
adressed_demand(
  baci,
  years = NULL,
  codes = NULL,
  year_ref,
  var_exporter,
  var_k,
  exporter_ref = NULL,
  base_100 = TRUE,
  compare = FALSE,
  return_output = TRUE,
  return_pq = FALSE,
  path_output = NULL
)
}
\arguments{
\item{baci}{Chemin d'accès, dataframe ou format parquet des données de baci
à utiliser.}

\item{years}{Années à garder dans les données.}

\item{codes}{Codes à garder dans les données.}

\item{year_ref}{Année de référence pour le calcul de la demande adressée.}

\item{var_exporter}{Variable contenant les exportateurs.}

\item{var_k}{Variable à utiliser pour le groupement des produits.}

\item{exporter_ref}{Exportateur de référence pour le calcul du ratio de la
demande adressée.}

\item{base_100}{Booléen indiquant si la demande adressée doit être calculée
en base 100 par rapport à l'année de référence.}

\item{compare}{Booléen indiquant si le ratio de la demande adressée doit être
calculé par rapport à un pays de référence.}

\item{return_output}{Booléen indiquant si le résultat doit être retourné.}

\item{return_pq}{Booléen indiquant si le résultat doit être retourné en format
parquet.}

\item{path_output}{Chemin d'accès pour enregistrer le résultat. Peut être
en format csv ou parquet.}
}
\value{
Un dataframe contenant la demande adressée.
}
\description{
Cacule la demande adressée à chaaue pays pour chaque année. La
demande adressée peut être calculée par pays ou par groupement de produits.
Le résultat retourné peut être la valeur de la demande adressée, la demande
adressée en base 100 ou bien le ratio de la abse 100 avec la base 100 d'un
pays de référence.
}
\details{
La demande adressée sert à observer l'évolution de la demande potentielle
adressée à un pays en prenant en compte le positionnement du pays sur
l'ensemble des marchés étudiés sur une année de référence. Elle est calculée
en sommant l'ensemble des importations de tous les pays sur un produit donné,
chaque valeur d'importation étant pondérée par la part que le pays représente
dans les exportations totale du pays étudié sur le produit donné.
La formule est la suivante :

\eqn{DA_{it} = \sum_{jk}M_{jkt} \times \frac{X_{ijkt=0}}{X_{it=0}}}
}
\examples{
# Pas d'exemples.
}
