% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_product.R
\name{ask_install_concordance}
\alias{ask_install_concordance}
\title{Ask to Install Concordance}
\usage{
ask_install_concordance(current_version = NULL)
}
\arguments{
\item{current_version}{The version of the concordance package. If NULL the
package is not installed.}
}
\description{
Prompts the user to install the appropriate version of the concordance package.
}
