% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_intern_data.R
\docType{data}
\name{chelem_classification}
\alias{chelem_classification}
\title{Geographic classification used by the database
\link{CHELEM} (\url{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=17}).}
\format{
\subsection{\code{classification_chelem}}{

A dataframe with 196 rows and 3 columns :
\describe{
\item{iso_region}{Str : Character iso code 3 use for the region.}
\item{name_region}{Str : The name of the region.}
\item{iso_country}{Str : Character iso code 3 use for the country}
}
}
}
\source{
\href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=17}{CHELEM classificiation}. See :
\href{http://www.cepii.fr/cepii/fr/publications/wp/abstract.asp?nodoc=1081}{de Saint Vaulry, A. (2008), “Base de données CHELEM - Commerce international du CEPII”,  Document de travail du CEPII, N°2008-09}
}
\usage{
chelem_classification
}
\description{
Geographic classification used by the database
\link{CHELEM} (\url{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=17}).
}
\keyword{datasets}
