% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_classification.R
\name{classification_berthou_2011}
\alias{classification_berthou_2011}
\title{Perform Flow Classification from Berthou & Emlinger 2011}
\usage{
classification_berthou_2011(df_baci, var_weighting, na.rm)
}
\arguments{
\item{df_baci}{BACI dataframe (arrow format)}

\item{var_weighting}{Variable to perform weighted geometric mean}

\item{na.rm}{Exclude NA or not}
}
\value{
Berthou 2011 classification
}
\description{
Perform Flow Classification from Berthou & Emlinger 2011
}
