% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_classification.R
\name{classification_fontagne_2007}
\alias{classification_fontagne_2007}
\title{Perform flow classification from Fontagne et al 2007}
\usage{
classification_fontagne_2007(df_baci, alpha, var_weighting, na.rm)
}
\arguments{
\item{df_baci}{BACI dataframe (arrow format)}

\item{alpha}{Numeric : Parameter for the threshold}

\item{var_weighting}{Character :  Variable to perform weighted median}

\item{na.rm}{Exclude NA or not}
}
\value{
BACI dataframe with fontagne et al 2007 classification
}
\description{
Perform flow classification from Fontagne et al 2007
}
