% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_classification.R
\name{classification_gaulier_2006}
\alias{classification_gaulier_2006}
\title{Perform flow classification from Gaulier et al 2006}
\usage{
classification_gaulier_2006(df_baci, weight, var_weighting = NULL, na.rm)
}
\arguments{
\item{df_baci}{BACI dataframe (arrow format)}

\item{weight}{Logical : Weighted quantiles not}

\item{var_weighting}{: Variable to perform weighted quantile if weight is TRUE}

\item{na.rm}{Exclude NA or not}
}
\value{
BACI dataframe with Gaulier et al 2006 classification
}
\description{
Perform flow classification from Gaulier et al 2006
}
