% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_intern_data.R
\docType{data}
\name{country_codes_V202401}
\alias{country_codes_V202401}
\title{Codes pays utilisés dans la base BACI + correspondance avec codes iso3}
\format{
\subsection{\code{country_codes_V202401}}{

Un dataframe avec 238 lignes et 2 colonnes :
\describe{
\item{country_code}{Codes pays utilisés dans la base BACI}
\item{country_iso3}{Correspondance avec les codes iso3}
}
}
}
\source{
\url{http://www.cepii.fr/DATA_DOWNLOAD/baci/data/BACI_HS92_V202401.zip}
}
\usage{
country_codes_V202401
}
\description{
Codes pays utilisés dans la base BACI + correspondance avec codes iso3
}
\keyword{datasets}
