% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_baci_db.R
\name{create_baci_db}
\alias{create_baci_db}
\title{Fonction qui permet de créer la databse BACI à partir des fichiers BACI et des critères choisis.}
\usage{
create_baci_db(
  baci_folder,
  year_start = NULL,
  year_end = NULL,
  product_code = NULL,
  exporter = NULL,
  importer = NULL,
  path_output = NULL,
  path_country_codes = NULL,
  return_df = TRUE
)
}
\arguments{
\item{baci_folder}{Chemin du dossier contenant les fichiers BACI (Ne modifier aucun nom par rapport à ce qui est téléchargé sur le site de la CEPII : http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37).}

\item{year_start}{Année de début de la base de données. Si non spécifiée, l'année la plus ancienne sera utilisée.}

\item{year_end}{Année de fin de la base de données. Si non spécifiée, l'année la plus récente sera utilisée.}

\item{product_code}{vecteur de codes HS6 utilisés pour filtrer le dataframe. Peut également être le dataframe créé par la fonction 'extract_product'.}

\item{exporter}{vecteur de numéros de pays (correspondance avec codes ISO3 trouvables dans le fichier country_codes.csv) utilisés pour filtrer le dataframe.}

\item{importer}{vecteur de numéros de pays (correspondance avec codes ISO3 trouvables dans le fichier country_codes.csv) utilisés pour filtrer le dataframe.}

\item{path_output}{Chemin où sauvegarder le dataframe. Si non spécifié, le dataframe ne sera pas sauvegardé.}

\item{path_country_codes}{Chemin du fichier country_codes.csv pour formater le dataframe. Si non spécifié, le dataframe ne sera pas formaté.}

\item{return_df}{Un booleen indiquant si l'on souhaite retourner le dataframe. Par defaut, return_df = TRUE.}
}
\value{
Un dataframe contenant les données de la base de données BACI importée et filtrée.
}
\description{
Fonction qui permet de créer la databse BACI à partir des fichiers BACI et des critères choisis.
}
\examples{
# Pas d'exemple pour cette fonction.
}
