% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quality_df.R
\name{create_quality_df}
\alias{create_quality_df}
\title{Création d'un dataframe servant à estimer la qualité d'un produit}
\source{
\href{http://www.cepii.fr/PDF_PUB/wp/2010/wp2010-23.pdf}{Gaulier, G. and Zignago, S. (2010) BACI: International Trade Database at the Product-Level. The 1994-2007 Version. CEPII Working Paper, N°2010-23}

\href{http://www.cepii.fr/DATA_DOWNLOAD/gravity/doc/Gravity_documentation.pdf}{Conte, M., P. Cotterlaz and T. Mayer (2022), "The CEPII Gravity database". CEPII Working Paper N°2022-05, July 2022.}

\href{https://www.sciencedirect.com/science/article/abs/pii/S0022199622000253?via\%3Dihub}{Fontagné L., Guimbard H. and Orefice G. (2022) Product-Level Trade Elasticities. Journal of International Economics, vol 137}
}
\usage{
create_quality_df(
  baci,
  gravity,
  years = NULL,
  codes = NULL,
  gravity_variables = NULL,
  revision_codes = "HS92",
  print = FALSE,
  return_output = TRUE,
  return_parquet = FALSE,
  path_output = NULL,
  format = "parquet",
  na.rm = TRUE
)
}
\arguments{
\item{baci}{Les données de BACI. Peut être un chemin vers un dossier parquet,
un dataframe ou un objet arrow.}

\item{gravity}{Les données de gravity. Peut être un chemin vers un dossier
parquet, un dataframe ou un objet arrow.}

\item{years}{Les années à garder. Par défaut, toutes les années sont gardées.}

\item{codes}{Les codes de produits à garder. Par défaut, tous les codes sont
gardés.}

\item{gravity_variables}{Les variables de gravité à garder. Par défaut, toutes
les variables sont gardées.}

\item{revision_codes}{La révision de laquelle les codes fournis proviennent.
Les valeurs possibles sont "HS92", "HS96", "HS02", "HS07", "HS12", "HS17" et
"HS22". Par défaut, les codes sont supposés être en HS92.}

\item{print}{Afficher le dataframe si TRUE. Par défaut, FALSE.}

\item{return_output}{Retourner le dataframe si TRUE. Par défaut, TRUE.}

\item{return_parquet}{Retourner le dataframe en format arrow si TRUE. Par
défaut, FALSE.}

\item{path_output}{Le chemin où sauvegarder le dataframe. Par défaut, NULL.}

\item{format}{Le format de sauvegarde du dataframe. Les valeurs possibles
sont "parquet" et "csv". Par défaut, "parquet".}

\item{na.rm}{Supprimer les valeurs manquantes si TRUE. Par défaut, TRUE.}
}
\value{
Un dataframe prêt à l'emploi pour estimer la qualité d'un produit.
}
\description{
Cette fonction permet de combiner les données des bases
\href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}{BACI},
\href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=8}{Gravity} et
\href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=35}{Product Level Trade Elasticities}
du \href{http://www.cepii.fr/CEPII/en/welcome.asp}{CEPII} afin d'obtenir un
dataframe prêt à l'emploi afin d'estimer (nottament) la qualité d'un produit
à partir d'une équation à la Khandelwal et al (2013).
}
\details{
Les données de BACI et de gravity étant volumineuses, ces dernières sont
traitées grâce au package arrow afin de limiter la prise de mémoire.
Cependant, un passage en mémoire est obligatoire à certains moments de
l'exécution de la fonction. Si les filtres (variables, années, codes) sont
trop permissifs comparés à la capacité de l'ordinateur, le processus peut
être fortement ralenti voir ne plus fonctionner. Dans ce cas, il est
recommandé de filtrer d'avantage les données et de procéder par itérations
successives.
}
\examples{
# Pas d'exemples.

}
