% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_baci.R
\name{dl_baci}
\alias{dl_baci}
\title{Download the BACI Database}
\source{
\href{http://www.cepii.fr/CEPII/fr/publications/wp/abstract.asp?NoDoc=2726}{Gaulier, G. et Zignago, S. (2010) BACI: International Trade Database at the Product-Level. The 1994-2007 Version. CEPII Working Paper, N°2010-23.}
}
\usage{
dl_baci(version = NULL, revision = "HS92", dl_folder)
}
\arguments{
\item{version}{A string indicating the version of BACI you wish to download.
By default it is set to NULL, which means that the latest available version
will be used.}

\item{revision}{A string indicating the revision of the Harmonised System in
which you want BACI. It must start with "HS" followed by a two digit number.
The default is "HS92", the oldest revision available.}

\item{dl_folder}{A string that specifies the folder in which you would like
to download the zip file of BACI.}
}
\value{
The BACI database in a zip file in the folder of your choice.
}
\description{
dl_baci' allows you to download the
\href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}{BACI}
produced by the \href{http://www.cepii.fr/CEPII/en/welcome.asp}{CEPII}. You can
select the revision and version of the BACI you require. The zip file will
be downloaded and saved to the location of your choice using the 'dl_folder'
parameter.
}
\examples{
# Download the latest version of the revision 1992 in your current directory
# dl_baci(dl_folder = here::here())

# Download the 202401 version of BACI for the 2007 revision
# dl_baci(version = "202401", revision = "HS07", dl_folder = here::here())

}
