% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_gravity.R
\name{dl_gravity}
\alias{dl_gravity}
\title{Download the Gravity Database}
\source{
Base \href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=8}{Gravity}
du \href{http://www.cepii.fr/CEPII/fr/welcome.asp}{CEPII}
}
\usage{
dl_gravity(dl_folder, dl_zip = FALSE)
}
\arguments{
\item{dl_folder}{Character indicating the path to the folder where the data
will be saved. Data will be saved in a sub-folder named "Gravity_csv_V(version)"
allowing to have more than one version.}

\item{dl_zip}{Logical indicating if the zip file must be dowloaded each time
even if it already exist (TRUE) or not (FALSE : the default). If FALSE, if
tjz ip file alreaqdy exist, only unzip and transform into parquet files.}
}
\value{
A folder containing the csv files and a folder containing the
parquet files of the Gravity Database.
}
\description{
Download the \href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=8}{Gravity}
Database created by the \href{http://www.cepii.fr/CEPII/en/welcome.asp}{CEPII} in
its last version the transform it into parquet files with one file by year.
}
\examples{
## dl_gravity(
##   dol_folder = "path-to save the data",
##   dl_zip = TRUE
## )

}
