% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameter.R
\name{.check_length_1}
\alias{.check_length_1}
\title{Check if Parameter is Length 1}
\usage{
.check_length_1(param, name_param)
}
\arguments{
\item{param}{Value of the parameter}

\item{name_param}{Character : name of the parameter}
}
\value{
Stop or nothing
}
\description{
Check if Parameter is Length 1
}
