% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameter.R
\name{.check_null_character}
\alias{.check_null_character}
\title{Check if Parameter is NULL or Character}
\usage{
.check_null_character(param, name_param)
}
\arguments{
\item{param}{Value of the parameter}

\item{name_param}{Character : name of the parameter}
}
\value{
Stop or nothing
}
\description{
Check if Parameter is NULL or Character
}
