% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameter.R
\name{.check_var_exist}
\alias{.check_var_exist}
\title{Check if Columns Provided are Present in a Dataframe}
\usage{
.check_var_exist(df, name_df, columns)
}
\arguments{
\item{df}{Dataframe}

\item{name_df}{Name of the Dataframe (of the parameter)}

\item{columns}{Character : Columns to be tested}
}
\value{
Stop or nothing
}
\description{
Check if Columns Provided are Present in a Dataframe
}
