% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-export_data.R
\name{.export_data}
\alias{.export_data}
\title{Export Data (Return and/or Save)}
\usage{
.export_data(
  data = NULL,
  return_output = TRUE,
  return_arrow = TRUE,
  path_output = NULL,
  eval,
  collect = NULL
)
}
\arguments{
\item{data}{Data to be export. Must be a R dataframe. Can be NULL if
\code{eval = FALSE} to just check the validity of the parameters.}

\item{return_output}{Logical indicating whether data must be returned or not.
By default data are returned after this function.}

\item{return_arrow}{Logical indicating whether data must be return in an
arrow format (TRUE the default) or not if \code{return_output = TRUE}.
By default data are returned to arrow format.}

\item{path_output}{Path to save the data. If NULL (default), the data
will not be saved. If \code{path_output} ends with a '.csv' extension, the data
will be saved in csv format. If no extension is given, the data will be
saved in a dataset parquet format in the specified folder. See the
\link[arrow]{arrow} package.}

\item{eval}{Logical indicating if the save/return part must be execute. If
FALSE, only the check validity or the parmaters will be performed.}

\item{collect}{A logical indicating if the data has been collected (TRUE) or
not (FALSE).}
}
\value{
Data saved and/or returned (or nothing).
}
\description{
Save data in parquet ore csv format and/or return data in
R dataframe format or \link{arrow} format.
}
\examples{
## Return data in arrow format. No save. Data had been collected
## .export_data(
##   data = df_data,
##   return_output = TRUE,
##   return_arrow = TRUE,
##   eval = TRUE
##   collect = TRUE
## )

## Return data in R dataframe format. Save in parquet format
## .export_data(
##   data = df_data,
##   return_output = TRUE,
##   return_arrow = FALSE,
##   path_output = "folder-data-parquet",
##   eval = TRUE,
##   collect = TRUE
## )

## Don't return data. Save in csv format. data had not been collected
## .export_data(
##   data = df_data,
##   return_output = FALSE,
##   return_arrow = FALSE,
##   path_output = "file-data.csv",
##   eval = TRUE,
##   collect = FALSE
## )

}
