% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_product.R
\name{ensure_devtools_installed}
\alias{ensure_devtools_installed}
\title{Ensure Devtools Installed}
\usage{
ensure_devtools_installed()
}
\description{
Checks if devtools is installed and installs it if not.
}
