% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_outliers_dist.R
\name{eval_outliers_dist}
\alias{eval_outliers_dist}
\title{Title}
\usage{
eval_outliers_dist(
  path_baci_parquet,
  years = NULL,
  codes = NULL,
  method = "classic",
  seuil_H_vector,
  seuil_L_vector,
  graph_type = "density",
  ref = "k",
  wrap = TRUE,
  output_type = "xlsx",
  path_output = NULL
)
}
\arguments{
\item{path_output}{}
}
\description{
Title
}
