% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_outliers_share.R
\name{eval_outliers_share}
\alias{eval_outliers_share}
\title{Evaluer le commerce restant après élimination des outliers}
\usage{
eval_outliers_share(
  path_baci_parquet,
  years = NULL,
  codes = NULL,
  method = "classic",
  seuil_H_vector,
  seuil_L_vector,
  graph = TRUE,
  path_df_output = NULL,
  path_graph_output = NULL
)
}
\arguments{
\item{path_baci_parquet}{Un chemin vers un dossier parquet contenant les
données BACI.}

\item{years}{Les années à considérer pour l'analyse. Si NULL, toutes les
années sont considérées.}

\item{codes}{Les codes HS6 à considérer pour l'analyse. Si NULL, tous les
codes HS6 sont considérés.}

\item{method}{Méthode de détermination des outliers. Peut être 'classic',
'fh13' ou 'h06'. Par défaut, 'classic'.}

\item{seuil_H_vector}{Un vecteur contenant les seuils "hauts" à considérer
pour l'élimination des outliers. Les valeurs doivent être comprises entre
0 et 1 pour les méthodes "classic" et "fh13".}

\item{seuil_L_vector}{Un vecteur contenant les seuils "bas" à considérer pour
l'élimination des outliers. Les valeurs doivent être comprises entre 0 et 1
pour les méthodes "classic" et "fh13".}

\item{graph}{Un booléen indiquant si un graphique doit être généré ou non.}

\item{path_df_output}{Un chemin vers un fichier csv où sauvegarder les
résultats de l'analyse. Si NULL, les résultats ne sont pas sauvegardés.}

\item{path_graph_output}{Un chemin vers un fichier où sauvegarder le
graphique généré. Si NULL, le graphique n'est pas sauvegardé.}
}
\value{
Un data.frame contenant les résultats de l'analyse et/ou un
graphique.
}
\description{
Permet de déterminr la part du commerce restante après élmination des
outliers pour plusieurs seuils différents.
}
\details{
La fonction permet de déterminer la part du commerce restante après
élimination des outliers pour plusieurs seuils différents.

La détermination des outliers se fait grâce à la fonction
\link{clean_uv_outliers}. Les valeurs et quantités
totales des parties de BACI sélectionnées sont calculées avant et après
élimination des outliers, ce qui permet de déterminer la part du commerce
aussi bien en valeur qu'en quantité restante après élimination des outliers.
Trois méthodes de détermination des outliers des outliers sont disponibles
et son documentées dans la fonction \link{clean_uv_outliers}.

Ce processus est effectué pour chaque seuil souhaité. les résultats sont
retournés sous forme de data.frame et/ou d'un graphique représentant les
données totales du commerce mais également les données par chapitre de
codes HS6 (les deux premiers chiffres du code HS6).
}
\examples{
# Pas d'exemple.
}
