% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_product.R
\name{extract_product}
\alias{extract_product}
\title{Fonction qui permet de creer un fichier xlsx ou csv contenant tous les codes hs6 souhaites ainsi que leur description.}
\usage{
extract_product(
  codes_vector,
  path_output,
  revision_origin = "HS22",
  revision_destination = "HS92",
  export = TRUE,
  return_df = TRUE
)
}
\arguments{
\item{codes_vector}{Un vecteur contenant les codes ou numeros de chapitres que l'on souhaite extraire.}

\item{path_output}{Chemin d'acces au fichier de sortie. Doit etre un fichier .xlsx ou .csv.}

\item{revision_origin}{Un caractere indiquant la revision des codes HS d'origine. Par defaut, revision_origin = "HS22". Les valeurs possibles sont "HS22", "HS92", "HS96", "HS02", "HS07", "HS12" et "HS17".}

\item{revision_destination}{Un caractere indiquant la revision des codes HS de destination. Par defaut, revision_destination = "HS92". Les valeurs possibles sont "HS92", "HS96", "HS02", "HS07", "HS12" et "HS17".}

\item{export}{Un booleen indiquant si l'on souhaite exporter le fichier. Par defaut, export = TRUE.}

\item{return_df}{Un booleen indiquant si l'on souhaite retourner le dataframe. Par defaut, return_df = TRUE.}
}
\value{
Un fichier .xlsx ou .csv contenant tous les codes HS6 souhaites ainsi que leur description.
}
\description{
Fonction qui permet de creer un fichier xlsx ou csv contenant tous les codes hs6 souhaites ainsi que leur description.
}
\examples{
# Pas d'exemple disponible.
}
