% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_baci.R
\name{filter_baci}
\alias{filter_baci}
\title{Filter the BACI Database}
\usage{
filter_baci(
  baci,
  years = NULL,
  codes = NULL,
  export_countries = NULL,
  import_countries = NULL
)
}
\arguments{
\item{baci}{Data to be loaded. Can be a path to a csv or an excel file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted.}

\item{years}{Numeric vector indicating the years to be kept in the variable \code{t}.}

\item{codes}{Character vector indicating the product codes to be kept in the
variable \code{k}.}

\item{export_countries}{Character or numeric vector indicating the
exporter countries to be kept in the variable \code{exporter} or \code{i}.}

\item{import_countries}{Character or numeric vector indicating the importer
countries to be kept in the variable \code{importer} or \code{j}.}
}
\value{
The BACI database filtered in an ArrowObject.
}
\description{
Keep only wanted data in the BACI database. You can filter by
years, product codes or by countries.
}
\details{
When filtering by countries ou can filter with the numeric iso codes
(variable i/j) or the character iso codes (variable exporter/importer).

It is recommended to use the function \code{\link[=dl_baci]{dl_baci}} to
obtain the BACI database. This way you are ensured to have good names and
types for the BACI variables.

The BACI database must have (if you are filtering by it) the following
variables :
\itemize{
\item \strong{t} : Numeric type. Used with \code{years}.
\item \strong{k} : Character type. Used with \code{codes}.
\item \strong{i} : Numeric type.  Used with \code{export_countries} with numeric inputs.
\item \strong{exporter} : Character type. Used with \code{export_countries} with character inputs.
\item \strong{j} : Numeric type. Used with \code{import_countries} with numeric inputs.
\item \strong{importer} : Character type. Used with \code{import_countries} with character inputs.
}

Data are loaded in ArrowObject with the function \code{\link[=load_data]{load_data}}.
}
\examples{
# Keep years from 2015 to 2022
# filter_baci(baci = "folder_parquet_files", years = c(2015:2022))

# Keep products 420211 and 420212
# filter_baci(baci = "folder_parquet_files", codes = c("420211", "420212"))

# Keep exporters France and Italy with iso codes character
# filter_baci(baci = "folder_parquet_files", export_countries = c("FRA", "ITA"))

# Keep exporters France and Italy with iso codes numeric
# filter_baci(baci = "folder_parquet_files", export_countries = c(251, 380))

}
