% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv_comp.R
\name{fixed_weight_computation}
\alias{fixed_weight_computation}
\title{Cretate Dataframe with Fixed Weight}
\usage{
fixed_weight_computation(
  df_baci,
  year_ref_fixed_weight,
  var_disaggregation,
  var_weight,
  var_temporal
)
}
\arguments{
\item{df_baci}{Dataframe.}

\item{year_ref_fixed_weight}{Numeric : Year to keep to the fixed weight.}

\item{var_disaggregation}{Vector of character indicating the variables
corresponding to the lower level of aggregation : generally
\code{var_disaggregation = c("exporter", "importer", "k", "t")}.}

\item{var_weight}{Name of the variable to be used for the ponderation.}

\item{var_temporal}{Name of the variable indicating the time.}
}
\value{
Dataframe with values in \code{var_pond} kept fixed.
}
\description{
Cretate Dataframe with Fixed Weight
}
