% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_classification.R
\name{flow_classification}
\alias{flow_classification}
\title{Variety Classification on BACI Data}
\source{
\href{http://cepii.fr/PDF_PUB/wp/1997/wp1997-07.pdf}{Fontagné, L., Freudenberg, M., & Péridy, N. (1997). Trade patterns inside the single market (No. 97-07). Paris: CEPII.}

\href{http://www.cepii.fr/PDF_PUB/wp/2006/wp2006-05.pdf}{G. Gaulier, F. Lemoine & D. Ünal-Kesenci (2006), “China's Emergence and the Reorganisation of Trade Flows in Asia”, CEPII Working Paper, n° 2006-05, March.}

\href{http://www.cepii.fr/PDF_PUB/wp/2007/wp2007-06.pdf}{L.Fontagné, G.Gaulier & S.Zignago (2007),”Specialisation across Varieties within Products and North-South Competition”,CEPII Working Paper, N°2007-06, May}

\href{http://www.cepii.fr/PDF_PUB/lettre/2011/let313.pdf}{A . Berthou, C . Emlinger (2011), « Les mauvaises performances françaises à l’exportation: La compétitivité prix est - elle coupable ? », La Lettre du CEPII , n°313, Septembre.}
}
\usage{
flow_classification(
  baci,
  years = NULL,
  codes = NULL,
  export_countries = NULL,
  import_countries = NULL,
  method = c("fontagne_1997", "gaulier_2006", "fontagne_2007", "berthou_2011"),
  alpha_H = NULL,
  alpha_L = alpha_H,
  var_weighting = NULL,
  na.rm = TRUE,
  weight = NULL,
  return_output = TRUE,
  return_arrow = TRUE,
  path_output = NULL
)
}
\arguments{
\item{baci}{BACI data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet.}

\item{years}{Numeric vector indicating the years to be kept in the variable \code{t}.}

\item{codes}{Character vector indicating the product codes to be kept in the
variable \code{k}.}

\item{export_countries}{Character or numeric vector indicating the
exporter countries to be kept in the variable \code{exporter} or \code{i}.}

\item{import_countries}{Character or numeric vector indicating the importer
countries to be kept in the variable \code{importer} or \code{j}.}

\item{method}{Character indicating the method to be used. it can be
\code{fontagne_1997} (the default), \code{gaulier_2006}, \code{fontagne_2007} or
\code{berthou_2011} (see details for more).}

\item{alpha_H}{Threshold used in \code{method = "fontagne_1997"} and
\code{method = "fontagne_2007"}.
Can be NULL if the selected method don't use it.}

\item{alpha_L}{Treshold used in \code{method = "fontagne_1997"}. By default it takes
the same value has \code{alpha_H}. Can be NULL if the selected method don't use it.}

\item{var_weighting}{A character indicating a numerical variable to be used
for weighting median, quantile or geometric mean. Can be NULL if
\code{method = "gaulier_2006"} and \code{weight = FALSE}.}

\item{na.rm}{Logical indicating if NA should be removed from the data. By
default it is set to TRUE. If FALSE you can obtain NA if NA are presents in
your data be carefull.}

\item{weight}{Logical indicating whether a weight should be applied. Used
only if \code{method = "gaulier_2006"}.}

\item{return_output}{Logical indicating whether data must be returned or not.
By default data are returned after this function.}

\item{return_arrow}{Logical indicating whether data must be return in an
arrow format (TRUE the default) or not if \code{return_output = TRUE}.
By default data are returned to arrow format.}

\item{path_output}{Path to save the data. If NULL (default), the data
will not be saved. If \code{path_output} ends with a '.csv' extension, the data
will be saved in csv format. If no extension is given, the data will be
saved in a dataset parquet format in the specified folder. See the
\link[arrow]{arrow} package.}
}
\value{
BACI data with appropriate variables depending on the choosen method
(see details for more).
}
\description{
Perform variety classification on the
\href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}{BACI}
database. 4 methods are available to classify each flow into a variety:
\itemize{
\item Fontagné, Freudenberg and Péridy (1997)
\item Gaulier, Lemoine and Ünal-Kesenci (2006)
\item Fontagné, Gaulier and Zignago (2007)
\item Berthou and Emlinger (2011)
}
}
\section{Classification methods}{
\subsection{Fontagné, Freudenberg and Péridy (1997)}{
\subsection{Explications}{

This method (available with \code{method = "fontagne_1997"}) classifies each trade
flow into a variety : high (H), medium (M) or low (L). For each trade flow,
the unit value (\eqn{v / q}) is calculated for each trade flow. For each
group year-product, the weighted median is calculated (usually weighted by
the value of the trade flow, but you can choose your variable with
\code{var_weighting}). To classify each flow into a variety, we
apply the following rule:
\itemize{
\item High if : \eqn{uv > (1 + \alpha_H) × median}

\item Low if : \eqn{uv < (1 / \alpha_L) × median}

\item Medium otherwise.
}

Generally \code{alpha_H} (\eqn{\alpha_H}) and \code{alpha_L} (\eqn{\alpha_L}) are set
to 1.15.
}

\subsection{requirement}{

To performing this method, the following parameters must be used : \code{baci},
\code{method = "fontagne_1997"}, \code{alpha_H}, \code{alpha_L} (by default is equal to
\code{alpha_H}), \code{var_weighting}, \code{na.rm}.

Your BACI data must have at minimum the following variables :
\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : the value of the trade flow}
\item{q}{Numeric : the quantity of the trade flow}
}

It is best to have your trade flows by countries
(ie : not aggregated by regions).
}

\subsection{Return}{

This method return at minimum the following variables :

\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : Value of the flow}
\item{q}{Numeric : Quantity of the flow}
\item{uv}{Numeric : Unit value of the flow}
\item{med_ref_t_k}{Numeric : Weighted median for each group t-k}
\item{fontagne_1997}{Character : Variety of the flow (H, M or L)}
\item{var_weighting}{Character : the variable used for the weigthingmedian}
\item{alpha_H}{Numeric : \eqn{\alpha_H}, threshold}
\item{alpha_L}{Numeric : \eqn{\alpha_L}, threshold}
}
}

}

\subsection{Gaulier, Lemoine and Ünal-Kesenci (2006)}{
\subsection{Explications}{

This method (available with \code{method = "gaulier_2006"}) classifies each trade
flow into a variety : high (H), medium (M) or low (L). For each trade flow,
the unit value (\eqn{v / q}) is calculated. For each group year-product, the
weighted (or not, depending on the \code{weight} parameter) 33th and 67th
quantiles are calculated. In general value is used to weight the quantile.
To classify each trade flow into a variety, we apply the following rule:
\itemize{
\item High if : \eqn{uv > q_{67}}
\item Low if : \eqn{uv < q_{33}}
\item Medium otherwise
}
}

\subsection{requirement}{

To perform this method, the following parameters must be used : \code{baci},
\code{weight}, \code{method = "gaulier_2006"}, \code{var_weighting}, \code{na.rm}.

Your BACI data must have at minimum the following variables :
\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : the value of the trade flow}
\item{q}{Numeric : the quantity of the trade flow}
}

It is best to have your flows by countries (ie : not aggregated by regions).
}

\subsection{Return}{

This method return at minimum the following variables :

\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : Value of the trade flow}
\item{q}{Numeric : Quantity of the trade flow}
\item{uv}{Numeric : Unit value of the trade flow}
\item{perc_33}{Numeric : 33th (Weighted) quantile for each group t-k}
\item{perc_67}{Numeric : 67th (Weighted) quantile for each group t-k}
\item{gaulier_2006}{Character : Variety of the trade flow (H, M or L)}
\item{ponderate}{Logical : TRUE if a weight is applied. FALSE otherwise}
\item{var_weighting}{Character : the variable used for the weigthing quantile
or NA is \code{weight = TRUE}}
\item{alpha_H}{Numeric : \eqn{\alpha_H}, threshold}
\item{alpha_L}{Numeric : \eqn{\alpha_L}, threshold}
}
}

}

\subsection{Fontagné, Gaulier and Zignago (2007)}{
\subsection{Explications}{

This method (available with \code{method = "fontagne_2007"}) makes it possible
to divide a trade flow into two varieties. A trade flow can have its value
split between Low (L) variety and Medium (M) variety, or between High (H)
variety and Medium (M) variety. A trade flow can also be classified as
Medium (M) only.

The unit value (\eqn{v / q}) is calculated for each trade flow. For each
group of year-product, the weighted median is calculated. In general, we
use the value of the trade flow to weight. A ratio \eqn{r = uv/uv_m} for
each trade flow is computed. Where \eqn{UV_{m}} is the weighted median for
the group year-product. Then a share (\eqn{S}) is calculated for each
variety according to the following rules:
\itemize{
\item if : \eqn{r < 1} : \eqn{S_L = 1 - r^{\alpha}} ; \eqn{S_M = r^{\alpha}}
\item If : \eqn{r > 1} : \eqn{S_H = 1 - 1/(r^{\alpha})} ; \eqn{S_M = 1/(r^{\alpha})}
\item If : \eqn{r = 1} : \eqn{S_M = 1}
}

with \eqn{\alpha} an exogeneous parameter fixed with \code{alpha_H}.
The more \eqn{\alpha} is small, the more \eqn{S_M} will be large.

Using these share we then calculate the value attributed to each variety
simply by multipliying the value of the trade flow and the share of the variety
for this trade flow.
}

\subsection{requirement}{

To perform this method, the following parameters must be used : \code{baci},
\code{method = "fontagne_2007"}, \code{alpha_H}, \code{var_weighting}, \code{na.rm}.

Your BACI data must have at minimum the following variables :
\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : the value of the trade flow}
\item{q}{Numeric : the quantity of the trade flow}
}

It is best to have your trade flows by countries (ie : not aggregated by regions).
}

\subsection{Return}{

This method return at minimum the following variables :

\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : Value of the trade flow}
\item{q}{Numeric : Quantity of the trade flow}
\item{uv}{Numeric : Unit value of the trade flow}
\item{med_ref_t_k}{Numeric : The weighted median for the group t-k}
\item{r}{Numeric : The ratio between \code{uv} and \code{med_ref_t_k}}
\item{share_L}{Numeric : The share attributed to the Low variety}
\item{share_M}{Numeric : The share attributed to the Medium variety}
\item{share_H}{Numeric : The share attributed to the High variety}
\item{v_L}{Numeric : The value of the trade flow attributed to the Low variety}
\item{v_M}{Numeric : The value of the trade flow attributed to the Medium variety}
\item{v_H}{Numeric : The value of the trade flow attributed to the High variety}
\item{var_weighting}{character : The variable used to weight}
}
}

}

\subsection{Berthou and Emlinger (2011)}{

This method (available with \code{method = "berthou_2011"}) classifies each trade
flow into a variety : low (L) or high (H). For each trade flow, the unit
value (\eqn{v / q}) is calculated. A weighted geometric mean is then
calculated for each year-product-importer group. Each tarde flow is
classified according to the following rule:
\itemize{
\item High if : \eqn{uv > geomean}
\item Low if : \eqn{uv <= geomean}
}
\subsection{requirement}{

To perform this method, the following parameters must be used : \code{baci},
\code{method = "berthou_2011"}, \code{var_weighting}, \code{na.rm}.

Your BACI data must have at minimum the following variables :
\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : the value of the trade flow}
\item{q}{Numeric : the quantity of the trade flow}
\item{j}{Numeric/character : Iso (numeric or character) of the importer}
}

It is best to have your tarde flows by countries (ie : not aggregated by regions).
}

\subsection{Return}{

This method return at minimum the following variables :

\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{v}{Numeric : Value of the trade flow}
\item{q}{Numeric : Quantity of the trade flow}
\item{j}{Numeric/character : Iso (numeric or character) of the importer}
\item{uv}{Numeric : Unit value of the trade flow}
\item{geomean_ref_t_k_j}{Numeric : The weighted geometric mean for the group t-k-j}
\item{berthou_2011}{character : The variety of the trade flow High (H) or Low (L)}
\item{var_weighting}{character : The variable used to weight}
}
}

}
}

\section{Computation}{
This feature uses \href{https://arrow.apache.org/docs/r/}{arrow} functionalities.
However, the computation of the various metrics must be in memory
(only this part is in memory). This can take some time depending on your
configuration and the size of your data. If the size of the data is too
large for your computer, it may crash. It is advisable to reduce the size
of your database and run this function several times.
}

\examples{
## fontagne_1997 method with both alpha set to 1.15, returned in arrow format
## flow_classification(
##   baci = "baci_folder_parquet",
##   method = "fontagne_1997",
##   alpha_H = 1.15,
##   alpha_L = 1.15,
##   var_weighting = "v",
##   na.rm = TRUE,
##   return_output = TRUE,
##   return_arrow = TRUE
## )

## gaulier_2006 method with weighting, save output in pq format and no return
## flow_classification(
##   baci = "baci-folder-parquet",
##   method = "gaulier_2006",
##   weight = TRUE
##   var_weighting = "v",
##   na.rm = TRUE,
##   return_output = FALSE,
##   return_arrow = FALSE,
##   path_output = "folder-output-gaulier-2006"
## )

## fontagne_2007 method with alpha = 3, save output in csv format and return R dataframe
## flow_classification(
##   baci = "baci-folder-parquet",
##   method = "fontagne_2007",
##   alpha_H = 3,
##   var_weighting = "v",
##   na.rm = TRUE,
##   return_output = TRUE,
##   return_arrow = FALSE,
##   path_output = "file-output-fontagne-2007.csv"
## )

## berthou_2011 method with alpha = 3, save output in pq format and return R dataframe
## flow_classification(
##   baci = "baci-folder-parquet",
##   method = "berthou_2011",
##   var_weighting = "v",
##   na.rm = TRUE,
##   return_output = TRUE,
##   return_arrow = FALSE,
##   path_output = "folder-output-berthou-2011.csv"
## )

}
\seealso{
\code{\link[=.load_data]{.load_data()}} For more informations concerning the loading.
\code{\link[=.filter_baci]{.filter_baci()}} For more informations concerning the filter of data inside the function.
\code{\link[=.export_data]{.export_data()}} For more informations concerning the export of the data inside the function.
}
