% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijkt_berthou_2011.R
\name{gamme_ijkt_berthou_2011}
\alias{gamme_ijkt_berthou_2011}
\title{Calcul des gammes de valeurs unitaires selon la méthode Berthou & Emlinger
(2011).}
\source{
\href{http://www.cepii.fr/PDF_PUB/lettre/2011/let313.pdf}{A . Berthou, C . Emlinger (2011), « Les mauvaises performances françaises à l’exportation: La compé titivité prix est - elle coupable ? », La Lettre du CEPII , n°313, Septembre.}
}
\usage{
gamme_ijkt_berthou_2011(
  path_baci_parquet,
  years = NULL,
  codes = NULL,
  return_output = FALSE,
  path_output = NULL,
  remove = FALSE
)
}
\arguments{
\item{path_baci_parquet}{Chemin d'accès vers le dossier où la base BACI est
stockée en format parquet.}

\item{years}{Les années à considérer (un vecteur de numériques). Par défaut,
toutes les années sont prises en compte.}

\item{codes}{Les codes des produits à considérer (un vecteur de chaînes
de caractères). Par défaut, tous les produits sont pris en compte.}

\item{return_output}{Un booléen qui permet de retourner le résultat de la
fonction. Par défaut, la fonction ne retourne pas de dataframe.}

\item{path_output}{Chemin vers le dossier où le résultat de la fonction doit
être stocké en format parquet par année. Par défaut, le résultat n'est pas
stocké.}

\item{remove}{Un booléen qui permet de supprimer tous les fichiers commençant
par t= dans le path_output s'il est non nul. Par défaut, FALSE.
Evite les confusions si plusieurs utilisations dans le même dossier.}
}
\value{
Un dataframe / dossier parquet contenant les données de la base BACI
avec le calcul des gammes. Les variables du dataframe sont les suivantes :
\describe{
\item{i}{Code iso numérique de l'importateur}
\item{j}{Code iso numérique de l'exportateur}
\item{k}{Code HS6 du produit (en chaîne de caractère)}
\item{t}{Année}
\item{v}{Valeur totale du flux en milliers de dollars courants}
\item{q}{Quantité du flux en tonne métrique}
\item{exporter}{Code iso3 de l'exportateur}
\item{importer}{Code iso3 de l'importateur}
\item{uv}{Valeur unitaire du flux en milliers dollars courants par
tonne métrique}
\item{geom_mean_weighted}{Moyenne géométrique pondérée par les valeurs des
valeurs unitaires des couples {t, k,j}}
\item{gamme_berthou_2011}{Gamme de valeur unitaire du flux commercial.
Peut être 'L' ou 'H'}
}
}
\description{
Détermine la gamme à laquelle appartient chaque flux commercial. Deux gammes
sont possibles : Low (L) et High (H). La gamme est déterminée en fonction de
la valeur unitaire du flux commercial. La valeur unitaire du flux commercial
est comparée à la moyenne géométrique pondérée des valeurs unitaires des
flux commerciaux.
}
\details{
Les gammes sont déterminées de la façon suivante : une moyenne géométrique
est calculée pour chaque couple {t, k, j} (année, produit, importateur).
Cette valeur indique la valeur unitaire moyenne des flux commerciaux pour
sur un produit donné pour un importateur donné sur une année donnée. La valeur
unitaire d'un flux commercial est comparée à cette moyenne géométrique. Si la
valeur unitaire est supérieure à la moyenne géométrique, le flux est classé
dans la gamme High (H). Sinon, le flux est classé dans la gamme Low (L).

Cette fonction utilise les fonctionnalité du package
\href{https://arrow.apache.org/docs/r/}{arrow} pour performer des calculs sans
avoir à charger BACI en mémoire. Cependant le calcul de la médiane
pondérée nécessite le passage de la base (uniquement la partie nécessaire)
en mémoire. Si la base est trop importante, les calculs peuvent prendre un
certain temps, voir entraîner un problème de mémoire de l'ordinateur.
Si cela arrive, il est conseillé de réduire le nombre d'années sur lesquelles
la fonction doit calculer les gammes et d'exécuter plusieurs fois la fonction
jusqu'à avoir toutes les années voulues.

Il est possible d'utiliser cette fonction sur un dossier parquet dans lequel
des calculs de gamme ont déjà été effectués avec d'autres méthodes. Cependant,
il est fortement déconseillé de le faire sur un dossier où les gammes de
prix ont été calculés à partir de la méthode de Fontagné et al. (2007), les
lignes de flux n'étant plus uniques.
}
\examples{
# Pas d'exemples.
}
