% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijkt_fontagne_1997.R
\name{gamme_ijkt_fontagne_1997}
\alias{gamme_ijkt_fontagne_1997}
\title{Fonction qui permet de déterminer la gamme des produits sur chaque marché (produit-pays) en utilisant la méthode développée par Fontagné et al (1997)}
\source{
Fontagné, L., Freudenberg, M., & Péridy, N. (1997). Trade patterns inside the single market (No. 97-07). Paris: CEPII.
}
\usage{
gamme_ijkt_fontagne_1997(
  path_baci_parquet,
  alpha_H = 0.15,
  alpha_L = alpha_H,
  years = NULL,
  codes = NULL,
  exporters = NULL,
  importers = NULL,
  return_output = FALSE,
  path_output = NULL,
  remove = FALSE
)
}
\arguments{
\item{path_baci_parquet}{Chemin vers le dossier où la base BACI est stockée en format parquet.}

\item{alpha_H}{Seuil pour déterminer les gammes hautes. Par défaut, 0.15 (uv > 1.15 * medf_ref).}

\item{alpha_L}{Seuil pour déterminer les gammes basses. Par défaut, 0.15 (uv < 1/0.85 * medf_ref).}

\item{years}{Les années à considérer (un vecteur de numériques). Par défaut, toutes les années sont prises en compte.}

\item{codes}{Les codes des produits à considérer (un vecteur de chaînes de caractères). Par défaut, tous les produits sont pris en compte.}

\item{exporters}{Les pays exportateurs à considérer (un vecteur de chaînes de caractères ou de numériques). Par défaut, tous les pays sont pris en compte.}

\item{importers}{Les pays importateurs à considérer (un vecteur de chaînes de caractères ou de numériques). Par défaut, tous les pays sont pris en compte.}

\item{return_output}{Un booléen qui permet de retourner le résultat de la fonction. Par défaut, la fonction ne retourne rien.}

\item{path_output}{Chemin vers le dossier où le résultat de la fonction doit être stocké en format parquet par année. Par défaut, le résultat n'est pas stocké.}

\item{remove}{Un booléen qui permet de supprimer tous les fichiers commençant par t= dans le path_output s'il est non nul. Par défaut, FALSE. Evite les confusions si plusieurs utilisations dans le même dossier.}
}
\value{
Un dataframe / dossier parquet contenant les données de la base BACI avec une colonne supplémentaire indiquant la gamme des produits sur chaque marché (produit-pays).
}
\description{
Fonction qui permet de déterminer la gamme des produits sur chaque marché (produit-pays) en utilisant la méthode développée par Fontagné et al (1997)
}
\examples{
# Pas d'exemples?
}
