% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijk_gaulier_2006.R
\name{gamme_ijkt_gaulier_2006}
\alias{gamme_ijkt_gaulier_2006}
\title{Fonction pour définir les gammes de prix selon la méthode de Gaulier et al (2006).}
\source{
G. Gaulier, F. Lemoine & D. Ünal-Kesenci (2006), “China's Emergence and the Reorganisation of Trade Flows in Asia”, CEPII Working Paper, n° 2006-05, March.
}
\usage{
gamme_ijkt_gaulier_2006(
  path_baci_parquet,
  pond = 1,
  years = NULL,
  codes = NULL,
  exporters = NULL,
  importers = NULL,
  return_output = FALSE,
  path_output = NULL
)
}
\arguments{
\item{path_baci_parquet}{Chemin vers le dossier où la base BACI est stockée en format parquet.}

\item{pond}{Un entier qui permet de définir la méthode de calcul des gammes de prix. Par défaut, pond = 1. Si pond = 1, les gammes sont calculées en utilisant les quantiles pondérés. Si pond = 2, les gammes sont calculées en utilisant les quantiles non pondérés. Si pond = 3, les gammes sont calculées en utilisant les quantiles pondérés et non pondérés.}

\item{years}{Les années à considérer (un vecteur de numériques). Par défaut, toutes les années sont prises en compte.}

\item{codes}{Les codes des produits à considérer (un vecteur de chaînes de caractères). Par défaut, tous les produits sont pris en compte.}

\item{exporters}{Les pays exportateurs à considérer (un vecteur de chaînes de caractères ou de numériques). Par défaut, tous les pays sont pris en compte.}

\item{importers}{Les pays importateurs à considérer (un vecteur de chaînes de caractères ou de numériques). Par défaut, tous les pays sont pris en compte.}

\item{return_output}{Un booléen qui permet de retourner le résultat de la fonction. Par défaut, la fonction ne retourne rien.}

\item{path_output}{Chemin vers le dossier où le résultat de la fonction doit être stocké en format parquet par année. Par défaut, le résultat n'est pas stocké.}
}
\value{
Un dataframe / dossier parquet contenant les données de la base BACI avec une colonne supplémentaire indiquant la gamme des produits sur chaque marché (produit-pays).
}
\description{
Fonction pour définir les gammes de prix selon la méthode de Gaulier et al (2006).
}
\examples{
# Pas d'exemples.
}
