% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lines_comparison.R
\name{graph_lines_comparison}
\alias{graph_lines_comparison}
\title{Create Line Graph}
\usage{
graph_lines_comparison(
  baci,
  x,
  y,
  linewidth = 1,
  var_linetype = NULL,
  manual_linetype = NULL,
  var_color = NULL,
  palette_color = NULL,
  manual_color = NULL,
  na.rm = TRUE,
  x_breaks = NULL,
  y_breaks = NULL,
  x_title = "",
  y_title = "",
  title = "",
  subtitle = "",
  caption = "",
  color_legend = "",
  linetype_legend = "",
  type_theme = c("bw", "classic", "minimal"),
  var_facet = NULL,
  path_output = NULL,
  width = 15,
  height = 8,
  print = TRUE,
  return_output = TRUE
)
}
\arguments{
\item{baci}{BACI data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet.}

\item{x}{Name of the variable used for the x axis. By default it is "t".}

\item{y}{Name of the variable used for the y axis.}

\item{linewidth}{Numeric indicating the width of the lines. By default it is
set to 1.}

\item{var_linetype}{Character indicating the name of the variable use
to define the type of the lines. By default it is set to NULL : all the
lines will be the same.}

\item{manual_linetype}{Character indicating the linetype to be used for each category.
Can be a vector of linetype or a list linking each linetype to a category.
By default this variable is NULL if no manual linetype is used. Used only
if \code{var_linetype = TRUE}.}

\item{var_color}{Name of the variable to be used to color the
lines. By default it is set to null indicating that all lines will have the
same default color.}

\item{palette_color}{Name of the palette to be used. See
\link[ggplot2]{scale_fill_brewer}. Be carefull, if the number of colors
in the palette is too small compared to the number of categories in
the variable \code{var_color}, an error will be returned.}

\item{manual_color}{Colors to be used for each category. Can be a vector
of colors or a list linking each color to a category. By default this variable
is NULL if no manual color is used.}

\item{na.rm}{Logical indicating whether NA should be exclude of the
graph (avoid warning message). TRUE is the default value.}

\item{x_breaks}{Values of the breaks on the x axis. By default the value is
NULL. In this case values are automatically choosen.}

\item{y_breaks}{Values of the breaks on the y axis. By default the value is
NULL. In this case values are automatically choosen.}

\item{x_title}{Title of the x axis. By default it is set to an empty string.}

\item{y_title}{Title of the y axis. By default it is set to an empty string.}

\item{title}{Title of the graph. By default it is set to an empty string.}

\item{subtitle}{Subtitle of the graph. By default it is set to an empty string.}

\item{caption}{Caption of the graph. By default it is set to an empty string.}

\item{color_legend}{Title of the legend for the colors. By default it is
set to "".}

\item{linetype_legend}{Title of the legend for the linetype. By default it
is set to "". if the variable for the linetype and the color is the same,
there will only be one legend.}

\item{type_theme}{General theme of the graph. It can be "bw" (the default)
\code{ggplot2::theme_bw()}, "classic" for \code{ggplot2::theme_classic()} or "minimal"
for \code{ggplot2::theme_minimal()}.}

\item{var_facet}{Name of the variable to be used to divide the graphs in
differents part with \code{ggplot2::facet_wrap()}. By default it his set on NULL
indicating that the graph will not be divide. Depends on the level of
aggregation of the data and the construction of the data.}

\item{path_output}{Path to save the data. If NULL (default), the data
will not be saved. If \code{path_output} ends with a '.csv' extension, the data
will be saved in csv format. If no extension is given, the data will be
saved in a dataset parquet format in the specified folder. See the
\link[arrow]{arrow} package.}

\item{width}{Width of the saved graph. Bu default it is set to 15}

\item{height}{Height of the saved graph. By default it is set to 8}

\item{print}{Logical indicating if the graph should be printed (default)
or not.}

\item{return_output}{Logical indicating whether data must be returned or not.
By default data are returned after this function.}
}
\value{
Graph representing the time evolution of a variable.
}
\description{
Represent a line graph allowing to represent the evolution of
a variable.
}
\section{General informations}{
This function don't modify data, so you must provide exact data you want
to represent.

Thanks to the \code{return_output} parameter, it is possible to return the graph,
allowing for further modifications.
}

\section{Theme default parameters}{
\subsection{Grid lines}{

Major and minor gridlines are removed (\code{panel.grid.minor()} and
\code{panel.grid.major()}).
}

\subsection{Titles options}{
\itemize{
\item The size of the title is 26 and it is centered.
\item the size of the subtitle is 22 and it is centered.
\item The size of the caption is 16 and it is left adjust.
}
}

\subsection{x axis options}{
\itemize{
\item The text of the x axis if at an angle of 45° and tight adjust. It has a
size of 18 and a "black" color.
\item The size of the x axis title is 22 and \code{vjust = -0.5}.
}
}

\subsection{y axis options}{
\itemize{
\item The size of the y axis text is 18 and its color is "black".
\item the size of the y axis title is 22.
}
}

\subsection{Legend options}{
\itemize{
\item legend position is right
\item Text of the legend has a size of 18 and a "black" color.
\item \code{legend.key.spacing.y = ggplot::unit(0.3, "cm")}.
\item the title of the legend has a size of 22, a "black" color and is centered.
}
}

\subsection{Facet options}{
\itemize{
\item The strip of the facet has a "black" colour and is fill with the following
color : \code{"#D9D9D9"}.
\item The text of the strip has a size of 18 and a "black" color.
}

By default facet are "free_y".
}
}

\examples{
## Basic example
## graph_lines_comparison(
##   baci = "data-folder-parquet",
##   x = "t",
##   y = "adressed-demand",
##   linewidth = 1,
##   var_linetype = "exporter",
##   var_color = exporter,
##   palette_color = "Paired",
##   type_theme = "minimal",
##   var_facet = "category"
## )

}
\seealso{
\code{\link[=.load_data]{.load_data()}} For more informations concerning the loading.
}
