% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_market_share.R
\name{graph_market_share}
\alias{graph_market_share}
\title{Graph of the Evolution of Market Share in Data}
\usage{
graph_market_share(
  baci,
  x = "t",
  y,
  graph_type = c("area", "line", "line_point"),
  var_fill_color,
  palette_color = NULL,
  manual_color = NULL,
  percent = TRUE,
  na.rm = TRUE,
  x_breaks = NULL,
  y_breaks = NULL,
  x_title = "",
  y_title = "",
  title = "",
  subtitle = "",
  caption = "",
  legend_title = "",
  type_theme = c("bw", "classic", "minimal"),
  var_facet = NULL,
  path_output = NULL,
  width = 15,
  height = 8,
  print = TRUE,
  return_output = TRUE
)
}
\arguments{
\item{baci}{BACI data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet.}

\item{x}{Name of the variable used for the x axis. By default it is "t".}

\item{y}{Name of the variable used for the y axis.}

\item{graph_type}{Type of the graph to be represented. It can be "area"
(by default and the suggested value) for an area graph, "line" for a line
graph or "line_point" for a line and point graph.}

\item{var_fill_color}{Name of the variable to be used to color the
graph.}

\item{palette_color}{Name of the palette to be used. See
\link[ggplot2]{scale_fill_brewer}. Be carefull, if the number of colors
in the palette is too small compared to the number of categories in
the variable \code{var_fill_color}, some categories an error will be returned.
By default this parameter is set on NULL if no palette color is used.}

\item{manual_color}{Colors to be used for each category. Can be a vector
of colors of a list linking each color to a category. By default this variable
is NULL if no manual color is used.}

\item{percent}{Logical indicating if labels on the y axis should be expressed
in percentage or not. TRUE is the default value. if TRUE data must already
be expressed in percentage.}

\item{na.rm}{Logical indicating whether NA should be exclude of the
graph (avoid warning message). TRUE is the default value.}

\item{x_breaks}{Values of the breaks on the x axis. By default the value is
NULL. In this case values are automatically choosen.}

\item{y_breaks}{Values of the breaks on the y axis. By default the value is
NULL. In this case values are automatically choosen.}

\item{x_title}{Title of the x axis. By default it is set to an empty string.}

\item{y_title}{Title of the y axis. By default it is set to an empty string.}

\item{title}{Title of the graph. By default it is set to an empty string.}

\item{subtitle}{Subtitle of the graph. By default it is set to an empty string.}

\item{caption}{Caption of the graph. By default it is set to an empty string.}

\item{legend_title}{Title of the color legend. By default it is set to an
empty string.}

\item{type_theme}{General theme of the graph. It can be "bw" (the default)
\code{ggplot2::theme_bw()}, "classic" for \code{ggplot2::theme_classic()} or "minimal"
for \code{ggplot2::theme_minimal()}.}

\item{var_facet}{Name of the variable to be used to divide the graphs in
differents part with \code{ggplot2::facet_wrap()}. By default it his set on NULL
indicating that the graph will not be divide. Depends on the level of
aggregation of the data and the construction of the data.}

\item{path_output}{Path to save the graph in "pdf" or "png" format. If NULL
(the default), the graph will not be saved.}

\item{width}{Width of the saved graph. Bu default it is set to 15}

\item{height}{Height of the saved graph. By default it is set to 8}

\item{print}{Logical indicating if the graph should be printed (default)
or not.}

\item{return_output}{Logical indicating if the graph must be returned for
further modifications (default) or not.}
}
\value{
Graph representing the evolution of market shares.
}
\description{
Create, export and save graph showing the evolution of the market share
(it can be other data, but this function has been think with the idea of
market share). Graph can be an area graph, a line graph or a line and point
graph. Some grahics paramters has been set for this function but all can
be change manually inside or outside the function (the graph can be
exported and so modified after the function).

This function does not calcuate the market share. Be sure to use data
a market share variable (or the variable you want to represent). It his
suggested to use the function \link{market_share} to compute the
market shares.
}
\section{Theme default parameters}{
\subsection{Grid lines}{

Major and minor gridlines are removed (\code{panel.grid.minor()} and
\code{panel.grid.major()}).
}

\subsection{Titles options}{
\itemize{
\item The size of the title is 26 and it is centered.
\item the size of the subtitle is 22 and it is centered.
\item The size of the caption is 16 and it is left adjust.
}
}

\subsection{x axis options}{
\itemize{
\item The text of the x axis if at an angle of 45° and tight adjust. It has a
size of 18 and a "black" color.
\item The size of the x axis title is 22 and \code{vjust = -0.5}.
}
}

\subsection{y axis options}{
\itemize{
\item The size of the y axis text is 18 and its color is "black".
\item the size of the y axis title is 22.
}
}

\subsection{Legend options}{
\itemize{
\item legend position is right
\item Text of the legend has a size of 18 and a "black" color.
\item \code{legend.key.spacing.y = ggplot::unit(0.3, "cm")}.
\item the title of the legend has a size of 22, a "black" color and is centered.
}
}

\subsection{Facet options}{
\itemize{
\item The strip of the facet has a "black" colour and is fill with the following
color : \code{"#D9D9D9"}.
\item The text of the strip has a size of 18 and a "black" color.
}

By default facet are "free_y".
}
}

\examples{
## graph area printed and saved
## market shares had been computed at the exporter level
## graph_market_share(
##   baci = "df-baci-market-share.csv",
##   x = "t",
##   y = "market_share_v",
##   graph_type = "area",
##   var_fill_color = "exporter",
##   palette_color = "Paired",
##   percent = TRUE,
##   title = "Title of the graph at the exporter level",
##   legend_title = "Exporters",
##   type_theme = "classic",
##   path_output = "path-to-png.png",
##   return_output = FALSE
## )

## graph area returned for modifications
## market shares had been computed at the region level
## One graph for each category of products
## graph_market_share(
##   baci = "df-baci-market-share.csv",
##   x = "t",
##   y = "market_share_v",
##   graph_type = "area",
##   var_fill_color = "region",
##   manual_color = c("blue", "red"),
##   percent = TRUE,
##   type_theme = "classic",
##   var_facet = "category",
##   return_output = TRUE,
##   print = FALSE
## )  +
##   ggplot2::theme(
##     legend.position = "left"
##   )

}
