% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_baci_file.R
\name{import_baci_file}
\alias{import_baci_file}
\title{Fonction qui permet d'importer un fichier BACI et de le filtrer selon les critères  choisis.}
\usage{
import_baci_file(
  path_baci_file,
  product_code = NULL,
  exporter = NULL,
  importer = NULL,
  path_country_codes = NULL
)
}
\arguments{
\item{path_baci_file}{Chemin du fichier BACI à importer.}

\item{product_code}{vecteur de codes HS6 utilisés pour filtrer le dataframe}

\item{exporter}{vecteur de numéros de pays (correspondance avec codes ISO3 trouvables dans le fichier country_codes.csv) utilisés pour filtrer le dataframe}

\item{importer}{vecteur de numéros de pays (correspondance avec codes ISO3 trouvables dans le fichier country_codes.csv) utilisés pour filtrer le dataframe.}

\item{path_country_codes}{Chemin du fichier country_codes.csv pour formater le dataframe. Si non spécifié, le dataframe ne sera pas formaté.}
}
\value{
Un dataframe contenant les données du fichier BACI importé et filtré.
}
\description{
Fonction qui permet d'importer un fichier BACI et de le filtrer selon les critères  choisis.
}
\examples{
# Pas d'exemple pour cette fonction.
}
