% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_product.R
\name{install_concordance_github}
\alias{install_concordance_github}
\title{Install Concordance from GitHub}
\usage{
install_concordance_github()
}
\description{
Installs the concordance package from GitHub.
}
