% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Khandelwal_quality_eq.R
\name{khandelwal_quality_eq}
\alias{khandelwal_quality_eq}
\title{Estimate Quality of a Trade Flow/Product (Khandelwal et al 2013)}
\source{
\href{https://www.aeaweb.org/articles?id=10.1257/aer.103.6.2169}{Khandelwal, A. K., Schott, P. K., & Wei, S. J. (2013). Trade liberalization and embedded institutional reform: Evidence from Chinese exporters. American Economic Review, 103(6), 2169-2195.}

\href{https://www.cairn.info/revue-notes-du-conseil-d-analyse-economique-2015-4-page-1.htm}{Bas, M., Fontagné, L., Martin, P., & Mayer, T. (2015). À la recherche des parts de marché perdues (Research Report No. 2015–23). Conseil d’Analyse Economique.}
}
\usage{
khandelwal_quality_eq(
  data_reg,
  reg_formula = NULL,
  y_var = NULL,
  x_var = NULL,
  fe_var = NULL,
  path_latex_output = NULL,
  title_latex = NULL,
  label_latex = NULL,
  print_reg_output = TRUE,
  return_output = TRUE
)
}
\arguments{
\item{data_reg}{Data to be used for the regression. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet. It is recommended to use the
\link{create_quality_df} to obtain these data.}

\item{reg_formula}{Formula object indicating the formula to be used for
the regression. If this parameter is non NULL, \verb{y_var, }x_var  et \code{fe_var}
will not be taking into account.}

\item{y_var}{Character indicating the name of the dependant variable
(classically it should be "demand").}

\item{x_var}{Character indicating the independant variables
(see details for more informations).}

\item{fe_var}{Character indictaing the fixed effects to be used
(See details for more informations).}

\item{path_latex_output}{Character indicating the path to where the
LaTeX table should be saved. If NULL table will not be saved.}

\item{title_latex}{Title of the LaTeX Table. By default NULL.}

\item{label_latex}{Label of the LaTeX table. By default NULL.}

\item{print_reg_output}{Logical indicating whether the results of the
regressions must be displayed (TRUE : the default) or not (FALSE).}

\item{return_output}{Logical indicating whether to return the output of
the rregression and the data containing the quality (TRUE the default) or not
(FALSE). If TRUE results are returned as a list with "lm" containing the
regression and "data_reg" the data.}
}
\value{
Results of the regression and data. Quality data are in level.
The following variables are added to the original data :
\describe{
\item{epsilon}{Numeric : residuals of the regression. If multiples
equations are estimated, it will be followed by the number of the regression
(epsilon_1, epsilon_2...)}
\item{quality}{Numeric : Level of the quality. If multiples
equations are estimated, it will be followed by the number of the regression
(quality_1, quality_2...)}
}
}
\description{
Allow to estimate the quality of a product with aggregated commercial
data like \href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}{BACI}.
The estimate equation can be freely choosen and can be multiple to
conduct robustness check. The creation of the data can be made
with the function \link{create_quality_df} from this package, and the
aggregation by prodict or sector can be made with the function
\link{aggregate_compare} from this package.
}
\details{
\subsection{Estimate equation}{

This function allow to estimate with a regression the "quality"
(non-price competitiveness) of a product based on a modelisation
proposed by Khandelwal et al (2013) and adapted to aggregated commercial
data by Bas et al (2015)

This modelisation is based on a demand function that depends on the price,
quality, the level of price and the wealth of the destination country.
The "quality" is defined as all the characteristics that increase the
demand for a same price.

It is estimate with the following regression :

\eqn{X_{ijkt} + \sigma_{k} p_{ijkt} = \beta PIB_{it} + \lambda D_{ij} + \alpha_{kjt} + \epsilon_{ijkt}}

X beeing the quantity of the trade flow; \eqn{\sigma_k} the elasticity of
the product ; p the price of the trade flow ; PIB the PIB of the
origin country ; D some gravity variables and \eqn{\alpha_{kjt}} a
product-importer-year fixed effect.

The logarithm of the non-price competitiveness is defined as the residuals
of this equation standardized by the elasticities of the product
\eqn{\frac{\epsilon_{ijkt}}{\sigma_{k} - 1}}. The level of the "quality" for
the flow is obtained by taking the exponential of this. The results of this
function is the level of "quality".
}

\subsection{Use of variables}{

This function use the function \code{\link[fixest]{feols}} of \pkg{fixest}
to estimate the regression.

To indicate the variables used for the regression, the user can
simply indicate the names of theses variables in a character vector in the
\code{x_var}, \code{y_var} and \code{fe_var} parameters. It is also possible to indicate
precisely the relation between the variable. By default it is simply an addition.
For example \code{x_var = c("var1", "var2")} is equivalent to \code{x_var = "var1 + var2"}.
For complex form it can be \code{x_var = "var1 * var2"}.

It is possible to add combined fixed effects with the symbol '^' between
the name of the fixed effects variables. For example \code{fe_var = c("var1", "var2")}
is equivalent to \code{fe_var = "var1 + var2"} whihc means two fixed effects.
To have only one combined fixed effect : \code{fe_var = "var1 ^ var2"}.

It is possible to estimate multiple regressions with stepwise functions :
see \code{\link[fixest]{feols}}. For example \code{var_x = "var1 + sw(var2, var3)"}
allow to estimate two regressions, one with \code{var1} and \code{var2} and the other
with \code{var1} and \code{var3}.

With the parmater \code{reg_formula}, you can directly give an object formula
to estimate the regression.
}
}
\examples{
## Estimate the regression and display regression table
## eq <- khandelwal_quality_eq(
##   data_reg = "path-to-parquet-folder",
##   y_var = "demand",
##   x_var = c("gdp_o", "contig", "dist") # variables taken from Gravity,
##   fe_var = "k^importer^t",
##   print_reg_output = TRUE,
##   return_output = TRUE
## )

## Take the data
## eq$data_reg

## Take the regression output
## eq$lm

}
