% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load the Data in Arrow Format}
\usage{
load_data(data)
}
\arguments{
\item{data}{Data to be loaded. Can be a path to a csv or an excel file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted.}
}
\value{
The data opened and loaded in an ArrowObject
}
\description{
Take data and load itin arrow format. Data can be a path to
a csv or an excel file. It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted.
}
