% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\name{market_share}
\alias{market_share}
\title{Compute Market Share}
\usage{
market_share(
  baci,
  years = NULL,
  codes = NULL,
  export_countries = NULL,
  import_countries = NULL,
  var_aggregate,
  var_share,
  na.rm = TRUE,
  return_output = TRUE,
  return_arrow = TRUE,
  path_output = NULL
)
}
\arguments{
\item{baci}{BACI data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet.}

\item{years}{Numeric vector indicating the years to be kept in the variable \code{t}.}

\item{codes}{Character vector indicating the product codes to be kept in the
variable \code{k}.}

\item{export_countries}{Character or numeric vector indicating the
exporter countries to be kept in the variable \code{exporter} or \code{i}.}

\item{import_countries}{Character or numeric vector indicating the importer
countries to be kept in the variable \code{importer} or \code{j}.}

\item{var_aggregate}{A character vector specifying the variables used to
aggregate the data in the calculation of market shares.

This parameter determines how the market share will be calculated by defining
the group of variables that will be used for aggregation.
\itemize{
\item \strong{Market Power} When there is only exporter or importer
(e.g, \code{c("t", "k", "exporter")}), the function calculates the market share
of the exporter for the product k in year t. This could reveal that
Country A export z\% of product k in year t.
}

If the aggregation is more detailed, the order of the variables in
\code{var_aggregate} is crucial as it changes the
perspective of the market share analysis:
\itemize{
\item \strong{Exporter Perspective}: When the exporter is listed first
(e.g., \code{c("t", "k", "exporter", "importer")}), the function calculates
the market share of the exporter within each importer. This allows you to
determine the proportion of a product exported
from a particular country to different importing countries. For example,
using this configuration could reveal that country A exports x\%
of its product k to country B.
\item \strong{Importer Perspective}: When the importer is listed first
(e.g., \code{c("t", "k", "importer", "exporter")}), the function calculates
the market share of the importer for products coming from each exporter.
This configuration helps in understanding the share of imports from specific
exporters into a particular importing country. For instance, this setup
could show that Country B imports y\% of its product k from Country C.
}

By adjusting the variables and their order in \code{var_aggregate}, you can
flexibly analyze market shares from different perspectives, providing
valuable insights into trade dynamics based on the specified aggregation.}

\item{var_share}{A character vector indicating variables to be used to
compute the market share. These variables must be numeric.}

\item{na.rm}{Logical indicating whether NA should be remove from the data for
the computation. By default it is set to TRUE. If FALSE you can obtain NA
if NA are presents in your data be carefull.}

\item{return_output}{Logical indicating whether data must be returned or not.
By default data are returned after this function.}

\item{return_arrow}{Logical indicating whether data must be return in an
arrow format (TRUE the default) or not if \code{return_output = TRUE}.
By default data are returned to arrow format.}

\item{path_output}{Path to save the data. If NULL (default), the data
will not be saved. If \code{path_output} ends with a '.csv' extension, the data
will be saved in csv format. If no extension is given, the data will be
saved in a dataset parquet format in the specified folder. See the
\link[arrow]{arrow} package.}
}
\value{
Dataframe with the following variables :
\describe{
\item{var_aggregate}{All the variables in var_aggregate}
\item{var_share}{All the variables in var_share}
\item{market_share_var_share}{the MArket share of each variable in
\code{var_share}. If \code{var_share = "v"}, the name will be \code{market_share_v}.}
}
}
\description{
Use the data to calculate the market share of the desired variables. The
level of of aggregation is freely selectable.
}
\details{
This function aggregate (sum) trade flow for each \code{var_aggregate} group for the
desired variables : \code{var_share}. Then shares are computed for each variables
in \code{var_share} with level of aggregation corresponding to \code{var_aggregate} minus
the last variable store in \code{var_aggregate}. Market shares are returned in
percentage.

This feature uses \href{https://arrow.apache.org/docs/r/}{arrow} functionalities
to load and filter the data. However, the computation of the must be in memory.
This can take some time depending on your
configuration and the size of your data. If the size of the data is too
large for your computer, it may crash. It is advisable to reduce the size
of your database and run this function several times.
}
\examples{
# Pas d'exemple.
## Share of v of the exporter for product 280530 and years 2020 to 2022
## market_share(
##   baci = "folder-baci-parquet",,
##   years = 2020:2022,
##   codes = "280530",
##   var_aggregate = c("t", "k", "exporter"), # Exporter in last is very important
##   var_share = "v",
##   return_arrow = FALSE  
## )

## Proportion of v and q exported from exporters to each importer
## Country A export x\% of product k to Country B
## market_share(
##   baci = "folder-baci-parquet",,
##   years = 2020:2022,
##   codes = "280530",
##   var_aggregate = c("t", "k", "exporter", "importer"), # Importer in last is very important
##   var_share = c("v", "q"),
##   return_arrow = FALSE  
## )

## Proportion of v and q imported from each exporters to an importer
## Country B import x\% of its product k from Country B
## market_share(
##   baci = "folder-baci-parquet",,
##   years = 2020:2022,
##   codes = "280530",
##   var_aggregate = c("t", "k", "importer", "exporter"), # Exporter in last is very important
##   var_share = c("v", "q"),
##   return_arrow = FALSE  
## )

}
\seealso{
\code{\link[=.load_data]{.load_data()}} For more informations concerning the loading.
\code{\link[=.filter_baci]{.filter_baci()}} For more informations concerning the filter of data inside the function.
\code{\link[=.export_data]{.export_data()}} For more informations concerning the export of the data inside the function.
}
