% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\name{market_share_variable}
\alias{market_share_variable}
\title{Compute the Market Share for the Desired Variable}
\usage{
market_share_variable(df, var_aggregate, var_share_indiv)
}
\arguments{
\item{df}{Dataframe}

\item{var_aggregate}{Character vector of variables to be used
for the aggregation of the market share}

\item{var_share_indiv}{Variable on which to compute the market share}
}
\value{
Dataframe with the market share computed
}
\description{
Compute the Market Share for the Desired Variable
}
