% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv_comp.R
\name{median_aggregation}
\alias{median_aggregation}
\title{Calculate the Median of a variable in a Dataframe}
\usage{
median_aggregation(df_baci, var, var_aggregation, na.rm)
}
\arguments{
\item{df_baci}{Baci data}

\item{var}{Variable to be aggregated}

\item{var_aggregation}{Character : vector of aggregations variables}

\item{na.rm}{Exclude NA or not}
}
\value{
Dataframe of aggregate unit values
}
\description{
Calculate the Median of a variable in a Dataframe
}
