% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_uv_outliers.R
\name{method_classic}
\alias{method_classic}
\title{Method Classic for Detecting Outliers}
\usage{
method_classic(df_baci, alpha_H, alpha_L, na.rm)
}
\arguments{
\item{df_baci}{BACI dataframe}

\item{alpha_H}{Threshold for high outliers}

\item{alpha_L}{Threshold for low outliers}

\item{na.rm}{Exclude NA or not}
}
\value{
BACI df with outliers Highlighted (R dataframe format)
}
\description{
Method Classic for Detecting Outliers
}
