% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_intern_data.R
\docType{data}
\name{product_codes_HS92_V202401}
\alias{product_codes_HS92_V202401}
\title{Codes HS6 de la nomenclature 1992 utilisés dans la base de donnée BACI sur la période 1995-2022}
\format{
\subsection{\code{product_codes_HS92_V202401}}{

Un dataframe avec 5022 lignes et 2 colonnes :
\describe{
\item{code}{Codes HS6 de la nomenclature 1992}
\item{description}{Description du produit}
}
}
}
\source{
\url{http://www.cepii.fr/DATA_DOWNLOAD/baci/data/BACI_HS92_V202401.zip}
}
\usage{
product_codes_HS92_V202401
}
\description{
Codes HS6 de la nomenclature 1992 utilisés dans la base de donnée BACI sur la période 1995-2022
}
\keyword{datasets}
