% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_intern_data.R
\docType{data}
\name{ptle}
\alias{ptle}
\title{Elasticités du commerce pour les produits HS6 : nomenclature HS2007}
\format{
\subsection{\code{ptle}}{

Un dataframe avec 5050 lignes et 4 colonnes :
\describe{
\item{HS6}{Codes HS6 de la nomenclature 2007}
\item{zero}{a dummy indicating whether the elasticity from the estimation was non-significant at the 1\% level}
\item{positive}{a dummy indicating whether the elasticity from the estimation was positive and significant}
\item{sigma}{the value of trade elasticity}
}
}
}
\source{
\url{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=35}

\href{https://www.sciencedirect.com/science/article/abs/pii/S0022199622000253?via\%3Dihub}{Fontagné L., Guimbard H. and Orefice G. (2022) Product-Level Trade Elasticities. Journal of International Economics, vol 137}
}
\usage{
ptle
}
\description{
Elasticités du commerce pour les produits HS6 : nomenclature HS2007
}
\keyword{datasets}
