% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_aggregate.R
\name{quality_aggregate}
\alias{quality_aggregate}
\title{Cacule la qualité aggrégé pour un ensemble de produits}
\usage{
quality_aggregate(
  data_quality,
  var_aggregate_k = "k",
  var_aggregate_i = "i",
  method_aggregate = "weighted.median",
  weighted_var = "q",
  fixed_weight = FALSE,
  year_ref = NULL,
  print_output = FALSE,
  return_output = TRUE,
  base_100 = FALSE,
  compare = FALSE,
  year_ref_base_100 = year_ref,
  exporter_ref = NULL,
  return_pq = FALSE,
  path_output = NULL
)
}
\arguments{
\item{data_quality}{Les données contenant les informations sur la qualité
de chaque flux. Peut-être un dataframe, un fichier parquet ou un objet
arrow. Il est recommandé d'utilisé les données renvoyées par la fonction
\link{khandelwal_quality_eq}.}

\item{var_aggregate_k}{La variable servant à aggréger les données. Par
défaut, la variable \code{k} est utilisée.}

\item{var_aggregate_i}{La variable des pays/régions d'origine servant à
aggréger les données. Par défaut la variable \code{i} est utilisée.}

\item{method_aggregate}{La méthode d'aggrégation des données. Par défaut,
la médiane pondérée avec \code{weighted.median} est utilisée. Les autres
méthodes disponibles sont : \code{mean}, \code{median} et
\code{weighted.mean}.}

\item{weighted_var}{La variable servant à pondérer les données. Par défaut,
la variable \code{q} est utilisée. Utile uniquement si la méthode
utilisée correspond à \code{weighted.mean} ou \code{weighted.median}.}

\item{fixed_weight}{Si \code{TRUE}, les pondérations sont fixées à une
année de référence. Par défaut, \code{FALSE}.}

\item{year_ref}{L'année de référence pour les pondérations. Utile uniquement
si \code{fixed_weight} est \code{TRUE}.}

\item{print_output}{Si \code{TRUE}, affiche les données aggrégées. Par
défaut, \code{FALSE}.}

\item{return_output}{Si \code{TRUE}, renvoie les données aggrégées. Par
défaut, \code{TRUE}.}

\item{base_100}{Booléen indiquant si la qualité doit être calculée en base
100 par rapport à une année de référence.}

\item{compare}{Booléen indiquant si la qualité en base 100 doit être comparée
à un pays de référence (par un ratio).}

\item{year_ref_base_100}{L'année de référence à utiliser si la qualité doit
être calculée en base 100.}

\item{exporter_ref}{L'entité exportatrice de référence à utiliser pour
la comparaison avec un pays de référence. L'entité doit se trouver dans la
variable \code{var_aggregate_i}.}

\item{return_pq}{Si \code{TRUE}, renvoie les données de qualité sous format
arrow. Par défaut, \code{FALSE}.}

\item{path_output}{Le chemin où sauvegarder les données aggrégées. Par
défaut, \code{NULL}. Peut être un chemin vers un fichier parquet ou un
fichier csv ou xlsx.}
}
\value{
Un dataframe contenant les données aggrégées de la qualité.
}
\description{
Calcule selon différentes manière une valeur aggrégée de la qualité pour
chaque produit/ensembles de produits.
}
\details{
Le calcul des données aggrégées de la qualité se fait obligatoirement
au niveau année-exporter. Il est possible de choisir sur quel ensemble de
produits, ces données sont calculées grâce au paramètre \code{var_aggregate_k}.

4 méthodes différentes sont disponibles pour effectuer l'aggrégations :
la moyenne arithmétique, la médiane, la moyenne arithmétique pondérée et la
médiane pondérée. Le choix de la variable de pondération s'effectue grâce au
paramètre \code{weighted_var}. Les pondérations peuvent varier chaque année,
ou bien être fixée à une année de référence grâce aux paramètres
\code{fixed_weight} et \code{year_ref}.

Pour utiliser correctement cette fonction, il faut que la fonction
\link{khandelwal_quality_eq} ait été utilisée au niveau des exportateurs et
non pas à un niveau plus agrégé de pays d'origine.
}
\examples{
# Pas d'exemple
}
