% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\name{sum_variable}
\alias{sum_variable}
\title{Sum values in the Desired Variable}
\usage{
sum_variable(df, var_aggregate, var_share_indiv, na.rm)
}
\arguments{
\item{df}{Dataframe}

\item{var_aggregate}{Character vector of variables to be used for
the aggregation of the sum}

\item{var_share_indiv}{Variable to sum}

\item{na.rm}{Exlude NA or not}
}
\value{
Dataframe with variable sum at the desired level
}
\description{
Sum values in the Desired Variable
}
