% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradalyze-package.R
\docType{package}
\name{tradalyze-package}
\alias{tradalyze}
\alias{tradalyze-package}
\title{tradalyze: Carry Out a Commercial Competitiveness Analysis}
\description{
Facilitate the analysis of trade competitiveness based on aggregated trade data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Xayoux/tradalyze}
  \item Report bugs at \url{https://github.com/Xayoux/tradalyze/issues}
}

}
\author{
\strong{Maintainer}: Romain Capliez \email{romain.capliez01@gmail.com}

}
\keyword{internal}
