% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfo_baci_pq.R
\name{transfo_baci_pq}
\alias{transfo_baci_pq}
\title{Transforme des fichiers csv de BACI en format parquet}
\usage{
transfo_baci_pq(csv_folder, path_output = csv_folder, version)
}
\arguments{
\item{csv_folder}{Chemin d'accès au dossier contenant les fichiers csv de BACI}

\item{path_output}{Chemin d'accès au dossier où seront stockés les fichiers parquet de BACI (par défaut, le même que csv_folder).
Ils seront stockés dans un dossier nommé "BACI-parquet".}

\item{version}{Version des fichiers BACI. Trouvable sur la page \href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}{BACI}}
}
\value{
Les fichiers parquet de BACI.
}
\description{
Transforme des fichiers csv de BACI en format parquet
}
\examples{
# Pas d'exemple.
}
