% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_gravity.R
\name{transfo_gravity_pq}
\alias{transfo_gravity_pq}
\title{Transforme La base de données Gravity en format parquet.}
\usage{
transfo_gravity_pq(csv_folder, path_output = csv_folder, version)
}
\arguments{
\item{csv_folder}{Chemin d'accès au dossier contenant les fichiers csv de
gravity}

\item{path_output}{Chemin d'accès au dossier où seront stockés les fichiers
parquet de Gravity (par défaut, le même que csv_folder).
Ils seront stockés dans un dossier nommé "Gravity-parquet".}

\item{version}{Version des fichiers Gravity. Trouvable sur la page
\href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=8}{Gravity}}
}
\value{
Les fichiers parquet de Gravity.
}
\description{
Transforme La base de données Gravity en format parquet.
}
\examples{
# Pas d'exemple.
}
