% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uv_comp.R
\name{uv_comp}
\alias{uv_comp}
\title{Comparaison des valeurs unitaires}
\usage{
uv_comp(
  baci,
  years = NULL,
  codes = NULL,
  formula = "median_pond",
  var_pond = NULL,
  year_ref,
  var_exporter,
  var_k,
  exporter_ref = NULL,
  base_100 = TRUE,
  compare = FALSE,
  return_output = TRUE,
  return_pq = FALSE,
  path_output = NULL
)
}
\arguments{
\item{baci}{Chemin d'accès, dataframe ou format parquet des données de baci
à utiliser.}

\item{years}{Années à garder dans les données.}

\item{codes}{Codes à garder dans les données.}

\item{formula}{Formule à utiliser pour calculer les valeurs unitaires. Peut
être "mean" pour la moyenne, "median" pour la médiane, "mean_pond" pour la
moyenne pondérée, ou "median_pond" pour la médiane pondérée.}

\item{var_pond}{Variable à utiliser pour la pondération des valeurs unitaire
lorsque formula demande une pondération.}

\item{year_ref}{Année de référence pour le calcul de la base 100.}

\item{var_exporter}{Variable contenant les exportateurs.}

\item{var_k}{Variable à utiliser pour le groupement des produits.}

\item{exporter_ref}{Exportateur de référence pour le calcul du ratio de la
base 100.}

\item{base_100}{Booléen indiquant si les valeurs unitaires doit être calculée
en base 100 par rapport à l'année de référence.}

\item{compare}{Booléen indiquant si le ratio des base 100 des valeurs
unitaires doit être calculé par rapport à un pays de référence.}

\item{return_output}{Booléen indiquant si le résultat doit être retourné.}

\item{return_pq}{Booléen indiquant si le résultat doit être retourné en format
parquet.}

\item{path_output}{Chemin d'accès pour enregistrer le résultat. Peut être
en format csv ou parquet.}
}
\value{
Un dataframe contenant l'évolution des valeurs unitaires.
}
\description{
Compare la moyenne des valeurs unitaires par produit ou groupe de produits
défini. La comparaison peut se faire en valeur, en base 100, ou en ratio par
par rapport à un pays de référence.
}
\details{
Un déterminant de l'évolution des exportations a trait à la compétitivité
prix, soit les prix fixés par les exportateurs par rapport à leurs
compétiteurs sur les marchés étrangers. Cette dimension combine les
évolutions de coût de production unitaire des entreprises, de leurs marges
et des évolutions du taux de change. Les valeurs unitaires du commerce
offrent une mesure de ces prix moyens à l'exportation.

Cette fonction prend comme input un dataframe ayant obligatoirement une
variable \code{t} contenant les années, ainsi qu'une variable \code{uv} contenant les
valeurs unitaires préalablement calculées. Ces variables sont directement
inclues dans les dataframes obtenus à partir des fonctions de ce package.
}
\examples{
# Pas d'exemples.
}
