% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_classification.R
\name{weighted_geomean}
\alias{weighted_geomean}
\title{Calculation of the Weighted Geometric Mean}
\source{
rogiersbart with his RTOOLZ package : https://rdrr.io/github/rogiersbart/rtoolz/man/weighted.geomean.html
}
\usage{
weighted_geomean(x, w, ...)
}
\arguments{
\item{x}{Numerical vector : data}

\item{w}{Numerical vector : weight}

\item{...}{Additional arguments}
}
\value{
Weighted geometric mean
}
\description{
Calculation of the Weighted Geometric Mean
}
