% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_chelem_classification.R
\name{add_chelem_classification}
\alias{add_chelem_classification}
\title{Ajoute la classification des régions du monde de CHELEM aux données de BACI}
\source{
\href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=17}{Classification Chelem du CEPII}. Voir :
\href{http://www.cepii.fr/cepii/fr/publications/wp/abstract.asp?nodoc=1081}{de Saint Vaulry, A. (2008), “Base de données CHELEM - Commerce international du CEPII”,  Document de travail du CEPII, N°2008-09}
}
\usage{
add_chelem_classification(
  baci,
  years = NULL,
  codes = NULL,
  path_output = NULL,
  return_output = FALSE,
  return_pq = FALSE
)
}
\arguments{
\item{baci}{Peut être un  chemin d'accès vers le dossier contenant
les données de BACI au format parquet. Peut également être un dataframe ou
bien des données au format arrow (requête ou non) permettant ainsi de chaîner
les opérations entre elles. ce paramètre est obligatoire.}

\item{years}{Années à conserver. Par défaut, toutes les années sont
conservées.}

\item{codes}{Codes à conserver. Par défaut, tous les codes sont conservés.}

\item{path_output}{Chemin d'accès vers le dossier où sauvegarder les données
fusionnées en format parquet. Par défaut, les données ne sont pas s
auvegardées.}

\item{return_output}{Retourner les données fusionnées. Par défaut, les données ne
sont pas retournées.}

\item{return_pq}{Retourner les données fusionnées au format parquet. Par
défaut, les données sont retournées au format tibble.}
}
\value{
Les données de BACI fusionnées avec la classification de CHELEM.
}
\description{
Ajoute la classification des régions du monde de CHELEM aux données de BACI.
Chaque exportateur et chaque importateurs se voient attribuer dans une région
géographique du monde à partir de la classification géographique utilisée par
la base de données \href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=17}{CHELEM}
du \href{http://www.cepii.fr/CEPII/fr/welcome.asp}{CEPII}.
}
\details{
Les données de la base de données \href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}{BACI}
sont exprimées au niveau des pays importateurs et exportateurs. Cependant,
il peut être souhaitable de regrouper les pays dans des zones géographiques
plus larges pour des analyses et résultats plus globaux.

Les données de BACI sont fusionnées avec la classification de CHELEM. Les
régions définies sont les suivantes :
\itemize{
\item North America
\item South America, central America and Caribbean
\item European Union -> supposée à 26 de manière constante (grande-Bretgane non
comprise).
\item CIS -> Commonwealth of Independent States
\item Others in Europe
\item North Africa
\item Sub-Sahara Africa
\item Near and Middle East
\item North-East Asia
\item South-East Asia
\item South Asia and Pacific
\item Australia and New Zealand
}

Les pays qui ne sont pas dans la classification de CHELEM sont attribués à la
région "rdm" pour "reste du monde".
}
\examples{
# Pas d'exemple
}
