% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adressed_demand.R
\name{adressed_demand}
\alias{adressed_demand}
\title{Compute the Adressed Demand}
\usage{
adressed_demand(
  baci,
  years = NULL,
  codes = NULL,
  export_countries = NULL,
  import_countries = NULL,
  year_ref,
  var_exporter,
  var_k,
  var_importer = "importer",
  base_100 = FALSE,
  compare = FALSE,
  exporter_ref = NULL,
  return_output = TRUE,
  return_arrow = TRUE,
  path_output = NULL
)
}
\arguments{
\item{baci}{BACI data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet.}

\item{years}{Numeric vector indicating the years to be kept in the variable \code{t}.}

\item{codes}{Character vector indicating the product codes to be kept in the
variable \code{k}.}

\item{export_countries}{Character or numeric vector indicating the
exporter countries to be kept in the variable \code{exporter} or \code{i}.}

\item{import_countries}{Character or numeric vector indicating the importer
countries to be kept in the variable \code{importer} or \code{j}.}

\item{year_ref}{Numeric indicating the year of reference to compute the
adressed demand and the base 100.}

\item{var_exporter}{Character indicating the name of the variable containing
the exporters.}

\item{var_k}{Character indicating the name of the variable containing
the products/group-products.}

\item{var_importer}{Character indicating the name of variable containing
the importers. it is recommended to let "importer".}

\item{base_100}{Logical indicating whether the adressed demand must be
transformed in base 100 (TRUE) or not (FALSE : the default). If TRUE two new
variables will be added :
\describe{
\item{DA_year_ref}{Numeric : Value of the adressed demand in the
reference year defined by \code{year_ref} parameter.}
\item{DA_100}{Numeric : Value of the adressed demand transformed
in base 100.}
}}

\item{compare}{Logical indicating if the base 100 should be compared (TRUE)
by a ratio, to the base 100 of a reference exporter, defined by the parameter
\code{exporter_ref}, or not (FALSE the default). If TRUE, two nex variables are
added
\describe{
\item{DA_100_exporter_ref}{Numeric : Value of the adressed demand for the
reference exporter defined by \code{exporter_ref} parameter.}
\item{DA_100_diff'}{Numeric : Value of the ratio between the base 100 and
the base 100 of the reference exporter. A value greater than 1 indicate that
the the variable for the exporter has increased more than the variable for
the reference exporter.}
}}

\item{exporter_ref}{Character indicating the reference exporter (or entity)
present in the variable given in \code{var_exporter}. Used only if
\code{compare = TRUE}. if not used it can be set on NULL.}

\item{return_output}{Logical indicating whether data must be returned or not.
By default data are returned after this function.}

\item{return_arrow}{Logical indicating whether data must be return in an
arrow format (TRUE the default) or not if \code{return_output = TRUE}.
By default data are returned to arrow format.}

\item{path_output}{Path to save the data. If NULL (default), the data
will not be saved. If \code{path_output} ends with a '.csv' extension, the data
will be saved in csv format. If no extension is given, the data will be
saved in a dataset parquet format in the specified folder. See the
\link[arrow]{arrow} package.}
}
\value{
BACI data with the following variables : variables contained in
\code{year_ref}, \code{var_exporter}, \code{var_k}, \code{var_importer}. There is also "DA" as new
variable. Depending on the value taken by \code{base_100} and
\code{compare} 2 or 4 more variables can be present. See the explications of these
parameters.
}
\description{
Compute the demand adress at each exporter, each year by product. The
result can be in level or in base 100.
}
\details{
The adress demand allow to observe the evolution of a potential demand
adressed to a country, taking into account the initial positionnement
of the country.

It is calculate by making the sum of the import for an importer and
weight it by the share this importer represent for the exporter. Then the
result for all importers is summed for each exporter.

\eqn{DA_{it} = \sum_{jk}M_{jkt} \times \frac{X_{ijkt=0}}{X_{it=0}}}

For this variable it is mandatory to have variables t and k in your dataframe
for the time and products.
}
\examples{
## Calculate the DA
## adressed_demand(
##   baci = "baci-parquet-folder",
##   year_ref = 2010,
##   var_exporter = "exporter",
##   var_k = "k",
##   base_100 = FALSE
## )

## Calculate the DA and take the base 100 and comapre it to France
## adressed_demand(
##   baci = "baci-parquet-folder",
##   year_ref = 2015,
##   var_exporter = "exporter",
##   var_k = "category",
##   base_100 = TRUE,
##   compare = TRUE,
##   exporter_ref = "FRA"
## )

}
\seealso{
\code{\link[=.load_data]{.load_data()}} For more informations concerning the loading.
\code{\link[=.filter_baci]{.filter_baci()}} For more informations concerning the filtering of data inside the function.
\code{\link[=.export_data]{.export_data()}} For more informations concerning the export of the data inside the function.
}
