% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chelem_classification.R
\docType{data}
\name{chelem_classification}
\alias{chelem_classification}
\title{Classification géographique utilisée par la base CHELEM au niveau de grandes régions}
\format{
\subsection{\code{classification_chelem}}{

Un dataframe avec 196 lignes et 3 colonnes :
\describe{
\item{iso_region}{Code iso3 utilisé pour définir les régions géographiques.}
\item{name_region}{Le nom associé à la région.}
\item{iso_country}{Code iso3 du pays entrant dans la composition de la région}
}
}
}
\source{
\href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=17}{Classification Chelem du CEPII}. Voir :
\href{http://www.cepii.fr/cepii/fr/publications/wp/abstract.asp?nodoc=1081}{de Saint Vaulry, A. (2008), “Base de données CHELEM - Commerce international du CEPII”,  Document de travail du CEPII, N°2008-09}
}
\usage{
chelem_classification
}
\description{
Classification géographique utilisée par la base CHELEM au niveau de grandes régions
}
\keyword{datasets}
