% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_uv_outliers.R
\name{clean_uv_outliers}
\alias{clean_uv_outliers}
\title{Déterminer et filtrer les outliers des valeurs unitaires dans BACI}
\source{
\href{https://pse.hal.science/hal-00959394/}{Lionel Fontagné, Sophie Hatte. European High-End Products in International Competition. 2013.}

\href{https://www.sciencedirect.com/science/article/abs/pii/S0022199605000516}{Hallak, J. C. (2006). Product quality and the direction of trade. Journal of international Economics, 68(1), 238-265.}
}
\usage{
clean_uv_outliers(
  path_baci_parquet,
  years = NULL,
  codes = NULL,
  method = "classic",
  seuil_H,
  seuil_L,
  whole = TRUE,
  visualisation = FALSE,
  path_output = NULL,
  return_output = FALSE
)
}
\arguments{
\item{path_baci_parquet}{Chemin vers le dossier BACI en format parquet.}

\item{years}{Années à garder dans les données. Si NULL, toutes les années
sont gardées.}

\item{codes}{Codes à garder dans les données. Si NULL, tous les codes sont
gardés.}

\item{method}{Méthode de détermination des outliers. Peut être 'classic',
'fh13' ou 'h06'.}

\item{seuil_H}{Seuil supérieur pour la détermination des outliers. Doit
être compris entre 0 et 1 pour les méthodes 'classic' et 'fh13'.}

\item{seuil_L}{Seuil inférieur pour la détermination des outliers. Doit
être compris entre 0 et 1 pour les méthodes 'classic' et 'fh13'.}

\item{whole}{Booléen pour déterminer si la distribution des différences
pour la méthode 'fh13' est celle de toutes les valeurs unitaires (TRUE) ou
celle des valeurs unitaires par produit (FALSE).}

\item{visualisation}{Booléen pour déterminer si les données exportées
doivent contenir les outliers ainsi que les variables pour les déterminer
ou non.}

\item{path_output}{Chemin vers le dossier où exporter les données en format
parquet. Si NULL, les données ne sont pas exportées.}

\item{return_output}{Booléen pour déterminer si les données doivent être
retournées en format data.frame.}
}
\value{
Un dataframe arrow contenant les données nettoyées des outliers.
}
\description{
Détermine les outliers dans la distribution des valeurs unitaires de BACI
et peut les retirer du jeu de données afin d'essayer d'éviter des biais
dans les mesures.
}
\details{
Les valeurs unitaires calculées à partir de BACI peuvent contenir des
outliers assez importants. En effet, les valeurs unitaire sont calculées
en divisant la valeur des exportations par la quantité exportée. Ces
dernières peuvent être sujettes à un certain nombre d'erreurs et leur
fiabilité n'est pas toujours garantie.

Pour éviter que ces valeurs ne biaisent les analyses, il peut être recommandé
de supprimer les outliers de la distribution des valeurs unitaires. Cette
fonction permet de déterminer les outliers selon trois méthodes différentes:
\itemize{
\item La méthode 'classic' détermine les outliers en fonction des quantiles de
la distribution des valuers unitaires par année-produit. Si la valeur
unitaire est supérieure au quantile déterminé par \code{seuil_H} ou inférieure
au quantile déterminé par \code{seuil_L}, alors elle est considérée comme un
outlier.
\item La méthode 'fh13' qui provient de l'article de
\href{https://pse.hal.science/hal-00959394/}{Fontagné & Hatte (2013)} détermine
les outliers en fonction de la différence entre la valeur unitaire et la
moyenne des valeurs unitaires par produit (de toute la période). Si la
différence est supérieure au quantile déterminé par \code{seuil_H} ou inférieure
au quantile déterminé par \code{seuil_L}, alors elle est considérée comme un
outlier. La distribution des différences peut être celle de toutes les
valeurs unitaires ou celle des valeurs unitaires par produit selon le
paramètre \code{whole}.
\item La méthode 'h06' qui provient de l'article de
\href{https://www.sciencedirect.com/science/article/abs/pii/S0022199605000516}{Hallak (2006)}
détermine les outliers en fonction de la moyenne des valeurs unitaires par
produit, exportateur et année. Si la valeur unitaire est supérieure à la
moyenne multipliée par \code{seuil_H} ou inférieure à la moyenne divisée par
\code{seuil_L}, alors elle est considérée comme un outlier.
}

La fonction permet de garder ou non les outliers dans les données.
}
\examples{
# Pas d'exemples.
}
