% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_baci_db.R
\name{create_baci_db}
\alias{create_baci_db}
\title{Fonction pour créer un dataframe/fichier csv à partir des données BACI, filtré selon les critères voulus.}
\usage{
create_baci_db(
  folder_baci,
  year_start = NULL,
  year_end = NULL,
  hs_codes = NULL,
  exporter_codes = NULL,
  importer_codes = NULL,
  add_iso3 = TRUE,
  calc_uv = TRUE,
  path_output = NULL,
  return_output = FALSE
)
}
\arguments{
\item{folder_baci}{Un chemin d'accès vers le dossier contenant les fichiers BACI (ne pas modifier les noms des fichiers par rapport aux noms originaux).}

\item{year_start}{Année de début de la base de données (Si NULL alors la date de départ sera l'année la plus ancienne des fichiers BACI).}

\item{year_end}{Année de fin de la base de données (Si NULL alors la date de fin sera l'année la plus récente des fichiers BACI).}

\item{hs_codes}{Un vecteur de caractères indiquant les codes HS à garder (Si NULL alors tous les codes HS sont gardés).}

\item{exporter_codes}{Un vecteur de numériques indiquant les codes pays exportateurs à garder (Si NULL alors tous les pays exportateurs sont gardés).}

\item{importer_codes}{Un vecteur de numériques indiquant les codes pays importateurs à garder (Si NULL alors tous les pays importateurs sont gardés).}

\item{add_iso3}{Un booléen indiquant si les codes iso3 doivent être ajoutés aux codes pays (TRUE par défaut).}

\item{calc_uv}{Un booléen indiquant si les valeurs unitaires doivent être calculées (TRUE par défaut).}

\item{path_output}{Un chemin d'accès vers le fichier csv de sortie (Si NULL alors aucun fichier n'est créé).}

\item{return_output}{Un booléen indiquant si le dataframe doit être retourné (FALSE par défaut). Attention si TRUE, le processus peut prendre très longtemps.}
}
\value{
Un dataframe et/ou un fichier csv selon les paramètres.
}
\description{
Fonction pour créer un dataframe/fichier csv à partir des données BACI, filtré selon les critères voulus.
}
\examples{
# Pas d'exemple pour cette fonction.
}
