% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_quality_df.R
\name{create_quality_df}
\alias{create_quality_df}
\title{Create a Dataframe Usefull for Estimate the Quality of a Product}
\source{
\href{http://www.cepii.fr/PDF_PUB/wp/2010/wp2010-23.pdf}{Gaulier, G. and Zignago, S. (2010) BACI: International Trade Database at the Product-Level. The 1994-2007 Version. CEPII Working Paper, N°2010-23}

\href{http://www.cepii.fr/DATA_DOWNLOAD/gravity/doc/Gravity_documentation.pdf}{Conte, M., P. Cotterlaz and T. Mayer (2022), "The CEPII Gravity database". CEPII Working Paper N°2022-05, July 2022.}

\href{https://www.sciencedirect.com/science/article/abs/pii/S0022199622000253?via\%3Dihub}{Fontagné L., Guimbard H. and Orefice G. (2022) Product-Level Trade Elasticities. Journal of International Economics, vol 137}
}
\usage{
create_quality_df(
  baci,
  gravity,
  years = NULL,
  codes = NULL,
  export_countries = NULL,
  import_countries = NULL,
  gravity_variables = NULL,
  baci_variables = NULL,
  revision_codes = c("HS0", "HS1", "HS2", "HS3", "HS4", "HS5", "HS6"),
  return_output = FALSE,
  return_arrow = FALSE,
  path_output = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{baci}{BACI data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet.}

\item{gravity}{Gravity data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because Gravity data are too large
and data must be in the first sheet.}

\item{years}{Numeric vector indicating the years to be kept in the variable \code{t}.}

\item{codes}{Character vector indicating the product codes to be kept in the
variable \code{k}.}

\item{export_countries}{Character or numeric vector indicating the
exporter countries to be kept in the variable \code{exporter} or \code{i}.}

\item{import_countries}{Character or numeric vector indicating the importer
countries to be kept in the variable \code{importer} or \code{j}.}

\item{gravity_variables}{Variables to keep from gravity in the output dataframe.
By default (if NULL) all the variables are beeing kept. No matter the value
of this parameter, the following variables are beeing kept :
exporter, importer, iso3_o, iso3_d.}

\item{baci_variables}{Variables to keep from baci in the output dataframe.
Bu default (if NULL) all the variables are beeing kept. No matter the value
of this parameter, the following variables are beeing kept :
exporter, importer, k, t, v et q.}

\item{revision_codes}{A character indicating the HS revision of the product
codes in BACI. By default this is the revision
"HS0" (1988/1992). The value can also be : "HS1" (1996), "HS2" (2022), "HS3" (2007),
"HS4" (2012) or "HS5" (2017). If the version of the
\link{concordance} package is greater than 2.1.0 you can also
choose "HS6" (2022).}

\item{return_output}{Logical indicating whether data must be returned or not.
By default data are returned after this function.}

\item{return_arrow}{Logical indicating whether data must be return in an
arrow format (TRUE the default) or not if \code{return_output = TRUE}.
By default data are returned to arrow format.}

\item{path_output}{Path to save the data. If NULL (default), the data
will not be saved. If \code{path_output} ends with a '.csv' extension, the data
will be saved in csv format. If no extension is given, the data will be
saved in a dataset parquet format in the specified folder. See the
\link[arrow]{arrow} package.}

\item{na.rm}{Logical indicating whether NA should be remove from the data for
the computation. By default it is set to TRUE. If FALSE you can obtain NA
if NA are presents in your data be carefull.}
}
\value{
Dataframe with the variable beeing kept from baci and gravity and the
following variables :
\describe{
\item{sigma}{Nuemric : Elasticities of the product taken from the PLTE database.}
\item{p}{Numeric : the "price" (unit value) of the trade flow (p = v / q)}
\item{demand}{Numeric : The value of the demand for this trade flow.
\eqn{demand = log(q) + (\sigma_k * log(p))}}
}
}
\description{
This function allow to combine data from
\href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}{BACI},
\href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=8}{Gravity}
and \href{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=35}{Product Level Trade Elasticities}
created by the \href{http://www.cepii.fr/CEPII/en/welcome.asp}{CEPII}. This
dataframe can be used to estimate the quality of a product
from an equation à la Khandelwal et al (2013) (see Bas et al 2015 for an
application with aggregate commercial data).
}
\details{
This feature uses \href{https://arrow.apache.org/docs/r/}{arrow} functionalities.
However, the computation of the various metrics must be in memory
(only this part is in memory). This can take some time depending on your
configuration and the size of your data. If the size of the data is too
large for your computer, it may crash. It is advisable to reduce the size
of your database and run this function several times.

The following variables are mandatory to have in baci data :
\describe{
\item{k}{Character : code product (or any other product indentifier)}
\item{t}{Numeric : Year}
\item{exporter}{Numeric/character : Iso numeric or character of the exporter}
\item{importer}{Numeric/character : Iso numeric or character of the importer}
\item{v}{Numeric : Value of the trade flow}
\item{q}{Numeric : Quantity of the trade flow}
}

The following variables are mandatory to have in Gravity data
\describe{
\item{iso3_o}{Character : Code iso numeric for the origin country.}
\item{iso3_d}{Character : Code iso numeric for the destination country.}
\item{year}{Numeric : Year.}
}

And of course all variables you want to have with \code{gravity_variables}
and \code{baci_variables}.
}
\examples{
## create_quality_df(
##   baci = "path-to-folder-baci-parquet",
##   gravity = "path-to-gravity-folder-parquet",
##   years = 2010:2020,
##   gravity_variables = c("dist", "contig", "comlang_off"),
##   revision_codes = "HS0",
##   return_output = TRUE,
##   return_parquet = FALSE,
##   path_output = "path-to-folder-output-parquet"
## )

}
\seealso{
\code{\link[=.load_data]{.load_data()}} For more informations concerning the loading.
\code{\link[=.filter_baci]{.filter_baci()}} For more informations concerning the filter of data inside the function.
\code{\link[=.export_data]{.export_data()}} For more informations concerning the export of the data inside the function.
}
