% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_gravity.R
\name{dl_gravity}
\alias{dl_gravity}
\title{Télécharge les données de la base de données Gravity}
\source{
Base \href{http://www.cepii.fr/CEPII/fr/bdd_modele/bdd_modele_item.asp?id=8}{Gravity}
du \href{http://www.cepii.fr/CEPII/fr/welcome.asp}{CEPII}
}
\usage{
dl_gravity(dl_folder, dl_zip = FALSE)
}
\arguments{
\item{dl_folder}{Chemin d'accès au dossier où seront stockés les fichiers
csv de Gravity. Ils seront stockés dans un dossier nommé
"Gravity_csv_V(version)". Ainsi si la version change un nouveau sous-dossier
sera créé.}

\item{dl_zip}{Si TRUE, télécharge le fichier zip de Gravity dans tous les cas.
Si FALSE, vérifie si le fichier zip existe déjà. Si oui, ne télécharge pas et
effectue uniquement le dezipage.}
}
\value{
Les fichiers parquet et csv de Gravity.
}
\description{
Télécharge la base de données Gravity depuis le site du CEPII dans sa
dernière version disponible puis la transforme en format parquet avec un
fichier par année.
}
\examples{
# Pas d'exemple
}
