% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_product.R
\name{extract_product}
\alias{extract_product}
\title{Fonction qui permet de creer un fichier xlsx ou csv contenant tous les codes hs6 souhaites ainsi que leur description.}
\usage{
extract_product(path_product_code, codes, path_output, export = TRUE)
}
\arguments{
\item{path_product_code}{Chemin d'acces au fichier contenant tous les codes produits HS6 (disponible dans le dossier BACI)}

\item{codes}{Un vecteur contenant les codes ou numeros de chapitres que l'on souhaite extraire.}

\item{path_output}{Chemin d'acces au fichier de sortie. Doit etre un fichier .xlsx ou .csv.}

\item{export}{Un booleen indiquant si l'on souhaite exporter le fichier. Par defaut, export = TRUE.}
}
\value{
Un fichier .xlsx ou .csv contenant tous les codes HS6 souhaites ainsi que leur description.
}
\description{
Fonction qui permet de creer un fichier xlsx ou csv contenant tous les codes hs6 souhaites ainsi que leur description.
}
\examples{
# Pas d'exemple disponible.
}
