% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijkt_berthou_2011.R
\name{gamme_ijkt_berthou_2011}
\alias{gamme_ijkt_berthou_2011}
\title{Fonction pour calculer les gammes de prix selon la méthode Berthou & Emlinger (2011).}
\usage{
gamme_ijkt_berthou_2011(
  path_baci_parquet,
  years = NULL,
  codes = NULL,
  exporters = NULL,
  importers = NULL,
  return_output = FALSE,
  path_output = NULL
)
}
\arguments{
\item{path_baci_parquet}{Chemin vers le dossier où la base BACI est stockée en format parquet.}

\item{years}{Les années à considérer (un vecteur de numériques). Par défaut, toutes les années sont prises en compte.}

\item{codes}{Les codes des produits à considérer (un vecteur de chaînes de caractères). Par défaut, tous les produits sont pris en compte.}

\item{exporters}{Les pays exportateurs à considérer (un vecteur de chaînes de caractères ou de numériques). Par défaut, tous les pays sont pris en compte.}

\item{importers}{Les pays importateurs à considérer (un vecteur de chaînes de caractères ou de numériques). Par défaut, tous les pays sont pris en compte.}

\item{return_output}{Un booléen qui permet de retourner le résultat de la fonction. Par défaut, la fonction ne retourne rien.}

\item{path_output}{Chemin vers le dossier où le résultat de la fonction doit être stocké en format parquet par année. Par défaut, le résultat n'est pas stocké.}
}
\value{
Un dataframe / dossier parquet contenant les données de la base BACI avec une colonne supplémentaire indiquant la gamme des produits sur chaque marché (produit-pays).
}
\description{
Fonction pour calculer les gammes de prix selon la méthode Berthou & Emlinger (2011).
}
\examples{
# Pas d'exemples.
}
