% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijkt_fontagne_1997.R
\name{gamme_ijkt_fontagne_1997}
\alias{gamme_ijkt_fontagne_1997}
\title{Calcul de gamme de valeurs unitaires selon la méthode de Fontagné, Freudenberg
et Péridy (1997)}
\source{
\href{http://cepii.fr/PDF_PUB/wp/1997/wp1997-07.pdf}{Fontagné, L., Freudenberg, M., & Péridy, N. (1997). Trade patterns inside the single market (No. 97-07). Paris: CEPII.}
}
\usage{
gamme_ijkt_fontagne_1997(
  path_baci_parquet,
  alpha_H = 1.15,
  alpha_L = alpha_H,
  years = NULL,
  codes = NULL,
  pivot = "longer",
  return_output = FALSE,
  path_output = NULL,
  remove = FALSE
)
}
\arguments{
\item{path_baci_parquet}{Chemin vers le dossier où la base BACI est stockée
en format parquet.}

\item{alpha_H}{Seuil pour déterminer les gammes hautes. Par défaut, 1.15
(uv > 1.15 * medf_ref). Peut être un vecteur de numériques.}

\item{alpha_L}{Seuil pour déterminer les gammes basses. Par défaut, 1.15
(uv < 1/0.85 * medf_ref). Peut être un vecteur de numériques.
Doit avoir le même nombre de valeurs que alpha_H.}

\item{years}{Les années à considérer (un vecteur de numériques). Par défaut,
toutes les années sont prises en compte.}

\item{codes}{Les codes des produits à considérer (un vecteur de chaînes de
caractères). Par défaut, tous les produits sont pris en compte.}

\item{pivot}{Un caractère qui indique si le format de sortie doit être
'longer' ou 'wider'. Par défaut, 'longer'. Un format longer indique que le
dataframe de sortie contient un variable pour le seuil H, une pour le seuil
L et une pour la gamme. Les observations sont égales au nombre d'observation
sélectionnées multipliées par le nombre de seuils. Un format wider indique
que le dataframe de sortie contient une variable par seuil et une observation
unique par flux.}

\item{return_output}{Un booléen qui permet de retourner le résultat de la
fonction. Par défaut, la fonction ne retourne rien.}

\item{path_output}{Chemin vers le dossier où le résultat de la fonction doit
être stocké en format parquet par année. Par défaut, le résultat n'est
pas stocké.}

\item{remove}{Un booléen qui permet de supprimer tous les fichiers commençant
par t= dans le path_output s'il est non nul. Par défaut, FALSE.
Evite les confusions si plusieurs utilisations dans le même dossier.}
}
\value{
Un dataframe / dossier parquet contenant les données de la base BACI
avec les gammes calculées.
\describe{
\item{i}{Code iso numérique de l'importateur}
\item{j}{Code iso numérique de l'exportateur}
\item{k}{Code HS6 du produit (en chaîne de caractère)}
\item{t}{Année}
\item{v}{Valeur totale du flux en milliers de dollars courants}
\item{q}{Quantité du flux en tonnes métriques}
\item{exporter}{Code iso3 de l'exportateur}
\item{importer}{Code iso3 de l'importateur}
\item{uv}{Valeur unitaire du flux en milliers de dollars courants/tonne
métrique}
\item{med_ref_t_k}{médiane pondérée par le commerce des valeurs unitaires
des flux pour chaque groupe (t, k)}
\item{gamme_fontagne_1997_...}{Gamme de la valeur unitaire du flux
(H, M, L)}
\item{seuil_H_}{Seuil pour déterminer les gammes hautes}
\item{seuil_L_}{Seuil pour déterminer les gammes basses}
}
}
\description{
Détermine la gamme à laquelle chaque flux appartient grâce à sa valeur
unitaire. Les trois gammes possibles sont Low (L), Medium (M) et High (H).
La méthode est basée sur la comparaison des valeurs unitaires avec la médiane
pondérée par la valeur commerciale des valeurs unitaires pour chaque marché
(produit-pays).
}
\details{
Les gammes sont déterminées de la façon suivante : la valeur unitaire de
chaque flux élémentaire est comparée à la médiane pondérée par la valeur
commerciale des valeurs unitaires pour chaque marché (produit-pays).
Si la valeur unitaire est supérieure à (1 + \code{alpha_H}) fois la médiane
pondérée, le flux est considéré comme High (H).
Si la valeur unitaire est inférieure à (1 / \code{alpha_L}) fois la médiane
pondérée, le flux est considéré comme Low (L). Sinon, le flux est
considéré comme Medium (M).

Selon cette méthodologie, les seuils basiques sont \code{alpha_H} = 1.15
et \code{alpha_L} = 1.15 (Un écart de 15\% à la médianne pondérée).

Cette fonction permet de performer le calcul de gammes pour différents
seuils de manière simultanée. Selon le paramètre \code{pivot}, le format de sortie
sera différent. En cas de format 'longer', le dataframe de sortie contiendra
une variable pour le seuil H, une pour le seuil L et une pour la gamme. Les
observations seront égales au nombre d'observation sélectionnées multipliées
par le nombre de seuils. Il faut donc faire attention à ne pas calculer
directement sur toutes les observations mais à bien trier selon les seuils
voulus. En cas de format 'wider', le dataframe de sortie contiendra une
variable gamme par seuil (avec les seuils indiqués dans le nom des variables)
et une observation unique par flux.

Cette dynamique des noms permet d'utiliser les données parquet créées pour
effectuer d'autres calculs de gammes (à l'exception des gammes de
fontagné et al 2007).

Cette fonction utilise les fonctionnalité du package
\href{https://arrow.apache.org/docs/r/}{arrow} pour performer des calculs sans
avoir à charger BACI en mémoire. Cependant le calcul de la médiane
pondérée nécessite le passage de la base (uniquement la partie nécessaire)
en mémoire. Si la base est trop importante, les calculs peuvent prendre un
certain temps, voir entraîner un problème de mémoire de l'ordinateur. Si
cela arrive, il est conseillé de réduire le nombre d'années sur lesquelles
la fonction doit calculer les gammes et d'exécuter plusieurs fois la fonction
jusqu'à avoir toutes les années voulues.
}
\examples{
# Pas d'exemples.
}
