% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijkt_fontagne_2007.R
\name{gamme_ijkt_fontagne_2007}
\alias{gamme_ijkt_fontagne_2007}
\title{Calcul de gamme selon la méthode de Fontagné, Gaulier & Zignago (2007)}
\source{
\href{http://www.cepii.fr/PDF_PUB/wp/2007/wp2007-06.pdf}{L.Fontagné, G.Gaulier & S.Zignago (2007),”Specialisation across Varieties within Products and North-South Competition”,CEPII Working Paper, N°2007-06, May}
}
\usage{
gamme_ijkt_fontagne_2007(
  baci,
  alpha = 3,
  years = NULL,
  codes = NULL,
  return_output = FALSE,
  path_output = NULL,
  remove = FALSE,
  return_pq = FALSE
)
}
\arguments{
\item{baci}{Peut être un  chemin d'accès vers le dossier contenant
les données de BACI au format parquet. Peut également être un dataframe ou
bien des données au format arrow (requête ou non) permettant ainsi de chaîner
les opérations entre elles. ce paramètre est obligatoire.}

\item{alpha}{Valeur de alpha pour le calcul des gammes (paramètre qui va
faire varier la répartition entre les gammes)}

\item{years}{Années à garder dans la base. Par défaut, toutes les années
sont gardées.}

\item{codes}{Codes à garder dans la base. Par défaut, tous les codes sont
gardés. Code doit être une chaîne de caractères.}

\item{return_output}{Un booléen qui permet de retourner le résultat de la
fonction. Par défaut, FALSE.}

\item{path_output}{Chemin vers le dossier où le résultat de la fonction doit
être stocké en format parquet par année. Par défaut, NULL.}

\item{remove}{Un booléen qui permet de supprimer tous les fichiers commençant
par t= dans le path_output s'il est non nul. Par défaut, FALSE.
Evite les confusions si plusieurs utilisations dans le même dossier.}

\item{return_pq}{Booléen pour indiquer si les données doivent être retournées
en format arrow si TRUE. Par défaut : FALSE.}
}
\value{
Un dataframe / dossier parquet contenant les données de la base
BACI avec les gammes calculées
\describe{
\item{i}{Code iso numérique de l'importateur}
\item{j}{Code iso numérique de l'exportateur}
\item{k}{Code HS6 du produit (en chaîne de caractère)}
\item{t}{Année}
\item{v}{Valeur totale du flux en milliers de dollars courants}
\item{q}{Quantité du flux en tonnes métriques}
\item{exporter}{Code iso3 de l'exportateur}
\item{importer}{Code iso3 de l'importateur}
\item{uv}{Valeur unitaire du flux en milliers de dollars courants par
tonne métrique}
\item{alpha_...}{Valeur du \eqn{\alpha} utilisé pour le calcul
des gammes}
\item{med_ref_ijkt}{Médiane pondérée par les valeurs unitaires des flux
pour chaque groupe (t, k)}
\item{r}{Ratio \eqn{r = UV_s/UV_m}}
\item{share_L_...}{Part du flux s dans la gamme Low. Si négatif,
le flux n'est pas dans la gamme}
\item{share_M_...}{Part du flux s dans la gamme Medium. Si négatif,
le flux n'est pas dans la gamme}
\item{share_H_...}{Part du flux s dans la gamme High. Si négatif,
le flux n'est pas dans la gamme}
\item{v_...}{Valeur du flux s dans la gamme correspondante en milliers
de dollars courants}
\item{gamme_fontagne_2007_...}{Gamme de valeur unitaire du flux s}
}
}
\description{
Calcul des gammes selon la méthode de Fontagné, Gaulier & Zignago (2007).
Chaque flux est divisé en une ou deux lignes selon sa/ses gammes.
Avec cette méthode il n'est pas conseillé d'effectuer des calculs de gammes
sur un dossier parquet possédant d'autres calculs de gammes, le nombre de
lignes n'étant pas le même.
}
\details{
Les gammes sont définies de la façon suivante : un ratio r est calculé pour
chaque flux s de la façon suivante :
\eqn{r = UV_s/UV_m} avec \eqn{UV_s} la valeur unitaire du flux s et
\eqn{UV_{m}} la médiane pondérée  par les valeurs commerciales des valeurs
unitaires des flux du groupe {t, k}.
\itemize{
\item Si \eqn{r < 1}, alors la part du flux s dans la gamme Low est
\eqn{1 - r^{\alpha}} et la part du flux s dans la gamme Medium est
\eqn{r^{\alpha}}.
\item Si \eqn{r > 1}, alors la part du flux s dans la gamme High est
\eqn{1 - 1/r^{\alpha}} et la part du flux s dans la gamme Medium est
\eqn{1/r^{\alpha}}.
\item Si \eqn{r = 1}, alors la part du flux s dans la gamme Medium est 1.
}

\eqn{\alpha} est un paramètre exogène qui va faire varier la répartition
entre les gammes. Plus il est petit, plus la part du commerce dans la gamme
moyenne sera grande.

Les flux élémentaires ne sont pas forcément classés dans une uniquement
gamme. Un flux peut donc correspondre à une ou deux lignes. Dans ce cas il
faut bien utiliser la variable v_{alpha} qui représente la valeur du flux
dans la gamme correspondante et non la variable v qui représente la valeur
totale du flux.

Cette fonction utilise les fonctionnalité du package
\href{https://arrow.apache.org/docs/r/}{arrow} pour ne pas charger les données
en mémoire. Cependant, la fonction doit mettre en mémoire la base pour
calculer les médianes pondérées avant de la repasser en format arrow.
Si la base sur laquelle les gammes sont calculées est trop imposante pour
l'ordinateur, le calcul peut prendre du temps /
être interrompu pour cause de manque de mémoire. Dans ce cas, il est
conseillé de réduire le nombre d'années et de ré-exécuter plusieurs
fois la fonction pour arriver au bon nombre d'années.
}
\examples{
# Pas d'exemples.
}
