% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijk_gaulier_2006.R
\name{gamme_ijkt_gaulier_2006}
\alias{gamme_ijkt_gaulier_2006}
\title{Calcul des gammes de valeurs unitaires selon la méthode de Gaulier et al
(2006).}
\source{
\href{http://www.cepii.fr/PDF_PUB/wp/2006/wp2006-05.pdf}{G. Gaulier, F. Lemoine & D. Ünal-Kesenci (2006), “China's Emergence and the Reorganisation of Trade Flows in Asia”, CEPII Working Paper, n° 2006-05, March.}
}
\usage{
gamme_ijkt_gaulier_2006(
  baci,
  pond = 1,
  years = NULL,
  codes = NULL,
  return_output = FALSE,
  return_pq = FALSE,
  path_output = NULL,
  remove = FALSE
)
}
\arguments{
\item{baci}{Peut être un  chemin d'accès vers le dossier contenant
les données de BACI au format parquet. Peut également être un dataframe ou
bien des données au format arrow (requête ou non) permettant ainsi de chaîner
les opérations entre elles. ce paramètre est obligatoire.}

\item{pond}{Un entier qui permet de définir la méthode de calcul des gammes
de prix. Par défaut, pond = 1. Si pond = 1, les gammes sont calculées en
utilisant les quantiles pondérés. Si pond = 2, les gammes sont calculées en
utilisant les quantiles non pondérés. Si pond = 3, les gammes sont calculées
en utilisant les quantiles pondérés et non pondérés (une variable pour
chaque).}

\item{years}{Les années à considérer (un vecteur de numériques). Par défaut,
toutes les années sont prises en compte.}

\item{codes}{Les codes des produits à considérer (un vecteur de chaînes de
caractères). Par défaut, tous les produits sont pris en compte.}

\item{return_output}{Un booléen qui permet de retourner le résultat de la
fonction. Par défaut, la fonction ne retourne rien.}

\item{return_pq}{Booléen pour indiquer si les données doivent être retournées
en format arrow si TRUE. Par défaut : FALSE.}

\item{path_output}{Chemin vers le dossier où le résultat de la fonction doit
être stocké en format parquet par année. Par défaut, le résultat n'est pas
stocké.}

\item{remove}{Un booléen qui permet de supprimer tous les fichiers commençant
par t= dans le path_output s'il est non nul. Par défaut, FALSE.
Evite les confusions si plusieurs utilisations dans le même dossier.}
}
\value{
Un dataframe / dossier parquet contenant les données de la base BACI
et les gammes calculées. Les variables du dataframe sont les suivantes :
\describe{
\item{i}{Code iso numérique de l'importateur}
\item{j}{Code iso numérique de l'exportateur}
\item{k}{Code HS6 du produit (en chaîne de caractère)}
\item{t}{Année}
\item{v}{Valeur totale du flux en milliers de dollars courants}
\item{q}{Quantité du flux en tonne métrique}
\item{exporter}{Code iso3 de l'exportateur}
\item{importer}{Code iso3 de l'importateur}
\item{uv}{Valeur unitaire du flux en milliers dollars courants par
tonne métrique}
\item{percentile_33}{33e percentile non pondéré des valeurs unitaires}
\item{percentile_67}{67e percentile non pondéré des valeurs unitaires}
\item{percentile_33_pond}{33e percentile pondéré par le commerce des
valeurs unitaires}
\item{percentile_67_pond}{67e percentile pondéré par le commerce des
valeurs unitaires}
\item{gamme_gaulier_2006_pond}{Gamme de valeur unitaire du flux commercial,
selon les percentiles pondérés. Peut être 'L', 'M' ou 'H'}
\item{gamme_gaulier_2006}{Gamme de valeur unitaire du flux commercial,
selon les percentiles non pondérés. Peut être 'L', 'M' ou 'H'}
}
}
\description{
Cette fonction permet de calculer les gammes de valeurs unitaires des produits
en utilisant la méthode de Gaulier et al. (2006). Trois gammes sont définies :
Low (L), Medium (M) et High (H). La gamme est déterminée en fonction du
percentile de la valeur unitaire du flux commercial.
}
\details{
Les gammes sont déterminées de la façon suivante : Si la valeur unitaire du
flux commercial se trouve sous le 33e percentile des valeurs unitaires
(pondéré par le commerce ou non) pour les groupes (t, k), le flux est classé
dans la gamme Low (L). Si la valeur unitaire du flux commercial se trouve au
dessus du 67e percentile des valeurs unitaires (pondéré par le commerce ou non)
pour les groupes (t, k), le flux est classé dans la gamme High (H). Sinon, le
flux est classé dans la gamme Medium (M).

Cette fonction utilise les fonctionnalité du package
\href{https://arrow.apache.org/docs/r/}{arrow} pour performer des calculs sans
avoir à charger BACI en mémoire. Cependant le calcul de la médiane
pondérée nécessite le passage de la base (uniquement la partie nécessaire)
en mémoire. Si la base est trop importante, les calculs peuvent prendre un
certain temps, voir entraîner un problème de mémoire de l'ordinateur.
Si cela arrive, il est conseillé de réduire le nombre d'années sur lesquelles
la fonction doit calculer les gammes et d'exécuter plusieurs fois la fonction
jusqu'à avoir toutes les années voulues.

Il est possible d'utiliser cette fonction sur un dossier parquet dans lequel
des calculs de gamme ont déjà été effectués avec d'autres méthodes. Cependant,
il est fortement déconseillé de le faire sur un dossier où les gammes de
prix ont été calculés à partir de la méthode de Fontagné et al. (2007), les
lignes de flux n'étant plus uniques.
}
\examples{
# Pas d'exemples.
}
