% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_bar_comp_year.R
\name{graph_bar_comp_year}
\alias{graph_bar_comp_year}
\title{Create Bar Graph to Compare Two Years}
\usage{
graph_bar_comp_year(
  baci,
  x,
  y,
  double_bar = FALSE,
  var_t,
  year_1,
  year_2,
  color_1 = FALSE,
  color_2 = FALSE,
  alpha = 0.7,
  var_fill = NULL,
  palette_fill = NULL,
  manual_fill = NULL,
  shape = 22,
  size_shape = 5,
  var_fill_shape = NULL,
  fill_shape = "black",
  na.rm = TRUE,
  x_title = "",
  y_title = "",
  title = "",
  subtitle = "",
  caption = "",
  legend_title = "",
  type_theme = c("bw", "classic", "minimal"),
  var_facet = NULL,
  path_output = NULL,
  width = 15,
  height = 8,
  print = TRUE,
  return_output = TRUE
)
}
\arguments{
\item{baci}{BACI data. Can be a path to a csv file.
It can also be a path to a folder containing
parquet files. Dataframe and ArrowObject are also accepted. Xlsx files are
also accepted, but absolutely not recommended because BACI data are too large
and data must be in the first sheet.}

\item{x}{Name of the variable used for the x axis. By default it is "t".}

\item{y}{Name of the variable used for the y axis.}

\item{double_bar}{Logical indicating whether bar must be stacked (TRUE), with
darker indicating the \code{year_1} and the clearer indicating the \code{year_2}, or
not (FALSE : the default value). In this case \code{year_2} is indicating by a shape.}

\item{var_t}{Character indicating the name of the temporal variable
from wich \code{year_1} and \code{year_2} are extracted.}

\item{year_1}{Numeric indicating the year 1.}

\item{year_2}{Numeric indicating the year 2.}

\item{color_1}{Character indicating the color of the \code{year_1} bar for the
outline. By default it is set to FALSE indicating that there will be no
outline.}

\item{color_2}{Character indicating the color of the \code{year_2} bar for the
outline. By default it is set to FALSE indicating that there will be no
outline. Not used if \code{doubel_bar = TRUE}.}

\item{alpha}{Numeric indicating the transparency of the \code{year_2} bar or
the shape representing \code{year_2} depending on \code{double_bar} parameter. By
default it is set to 0.7. It is not recommended to put it to 1.}

\item{var_fill}{Character indicating the variable to be used to color
the bars (the two bars in the case of \code{double_bar = TRUE} or the unique bar
if \code{double_bar = FALSE}). If NULL (the default) is provided, the basic color
will be used.}

\item{palette_fill}{Character indicating the name of the palette to be used. See
\link[ggplot2]{scale_fill_brewer}. Be carefull, if the number of colors
in the palette is too small compared to the number of categories in
the variable \code{var_fill_color}, an error will be returned.
By default this parameter is set on NULL if no palette color is used.}

\item{manual_fill}{Character indicating the colors to be used for each category.
Can be a vector of colors or a list linking each color to a category.
By default this variable is NULL if no manual color is used. These color
applicates also to the shape if \code{var_fill_shape} if used.}

\item{shape}{Numeric corresponding to the number of the shape wanted
if \code{double_bar = FALSE}. By default it is set to 22 (a full square). It is
recommended to use shape above or equal to 21 beacause these shape are full
and can be colored.}

\item{size_shape}{Numeric indicating the
size of the shape if \code{double_bar = TRUE}. By default it is set to 5.}

\item{var_fill_shape}{Character indicating the name of the variable
to be used to color the shape if \code{double_bar = TRUE}. It is recommended to
use the same as \code{var_fill}. If NULL (the default) is provided you can
use \code{fill_shape} to set a unique color.}

\item{fill_shape}{Character indicating the color to be used for the shape
if \code{double_bar = TRUE} and \code{var_fill_shape = NULL}. The color will be the
same for all shapes. The default is "black".}

\item{na.rm}{Logical indicating whether NA should be exclude of the
graph (avoid warning message). TRUE is the default value.}

\item{x_title}{Title of the x axis. By default it is set to an empty string.}

\item{y_title}{Title of the y axis. By default it is set to an empty string.}

\item{title}{Title of the graph. By default it is set to an empty string.}

\item{subtitle}{Subtitle of the graph. By default it is set to an empty string.}

\item{caption}{Caption of the graph. By default it is set to an empty string.}

\item{legend_title}{Title of the color legend. By default it is set to an
empty string.}

\item{type_theme}{General theme of the graph. It can be "bw" (the default)
\code{ggplot2::theme_bw()}, "classic" for \code{ggplot2::theme_classic()} or "minimal"
for \code{ggplot2::theme_minimal()}.}

\item{var_facet}{Name of the variable to be used to divide the graphs in
differents part with \code{ggplot2::facet_wrap()}. By default it his set on NULL
indicating that the graph will not be divide. Depends on the level of
aggregation of the data and the construction of the data.}

\item{path_output}{Path to save the data. If NULL (default), the data
will not be saved. If \code{path_output} ends with a '.csv' extension, the data
will be saved in csv format. If no extension is given, the data will be
saved in a dataset parquet format in the specified folder. See the
\link[arrow]{arrow} package.}

\item{width}{Width of the saved graph. Bu default it is set to 15}

\item{height}{Height of the saved graph. By default it is set to 8}

\item{print}{Logical indicating if the graph should be printed (default)
or not.}

\item{return_output}{Logical indicating whether data must be returned or not.
By default data are returned after this function.}
}
\value{
Graph representing the level and evolution of a variable for 2 years
}
\description{
Represent a bar graph allowing to compare the level and evolution of
different groups in a variable.
}
\section{Representations}{
There is two types of representation. The first is by stacking the bar
with \code{double_bar = TRUE}. The \code{year_1} is represented by a darker bar and the
\code{year_2} is represented by a clear bar above. Be carefull with this
representation because if the value of \code{year_1} is greter than the value
of \code{year_2} the clearer bar will be hidden by the darker bar and only the
outline will be visible if \code{color_2} is provided.

The second representation consist of one bar representing the \code{year_1} and
the \code{year_2} is representing by a shape.
}

\section{General informations}{
This function don't modify data, so you must provide exact data you want
to represent.

Thanks to the \code{return_output} parameter, it is possible to return the graph,
allowing for further modifications.
}

\section{Theme default parameters}{
\subsection{Grid lines}{

Major and minor gridlines are removed (\code{panel.grid.minor()} and
\code{panel.grid.major()}).
}

\subsection{Titles options}{
\itemize{
\item The size of the title is 26 and it is centered.
\item the size of the subtitle is 22 and it is centered.
\item The size of the caption is 16 and it is left adjust.
}
}

\subsection{x axis options}{
\itemize{
\item The text of the x axis if at an angle of 45° and tight adjust. It has a
size of 18 and a "black" color.
\item The size of the x axis title is 22 and \code{vjust = -0.5}.
}
}

\subsection{y axis options}{
\itemize{
\item The size of the y axis text is 18 and its color is "black".
\item the size of the y axis title is 22.
}
}

\subsection{Legend options}{
\itemize{
\item legend position is right
\item Text of the legend has a size of 18 and a "black" color.
\item \code{legend.key.spacing.y = ggplot::unit(0.3, "cm")}.
\item the title of the legend has a size of 22, a "black" color and is centered.
}
}

\subsection{Facet options}{
\itemize{
\item The strip of the facet has a "black" colour and is fill with the following
color : \code{"#D9D9D9"}.
\item The text of the strip has a size of 18 and a "black" color.
}

By default facet are "free_y".
}
}

\examples{
## Bar stacked
## graph_bar_comp_year(
##   baci = "data-folder-parquet",
##   x = "exporter",
##   y = "uv",
##   double_bar = TRUE,
##   var_t = "t",
##   year_1 = 2010,
##   year_2 = 2022,
##   color_1 = "black",
##   color_2 = "black",
##   alpha = 0.7,
##   var_fill = "exporter",
##   palette_fill = "Paired",
##   title = "Compare the level and evolution of unit values between 2010 and 2022",
##   type_theme = "bw",
##   var_facet = "sector"
## ) +
##   ggplot2::theme(
##     legend.position = "none" # No legend display
##   )

## Bar and shape
## graph_bar_comp_year(
##   baci = "data-folder-parquet",
##   x = "exporter",
##   y = "quality",
##   double_bar = FALSE,
##   var_t = "t",
##   year_1 = 2010,
##   year_2 = 2022,
##   alpha = 0.7,
##   var_fill = "exporter",
##   palette_fill = "Paired",
##   shape = 21,
##   size_shape = 6,
##   var_fill_shape = "exporter"
##   title = "Compare the level and evolution of unit values between 2010 and 2022",
##   type_theme = "classic"
## ) +
##   ggplot2::theme(
##     legend.position = "none" # No legend display
##   )

}
\seealso{
\code{\link[=.load_data]{.load_data()}} For more informations concerning the loading.
}
