% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lines_comparison.R
\name{graph_lines_comparison}
\alias{graph_lines_comparison}
\title{Graphique de la demande adressée.}
\usage{
graph_lines_comparison(
  baci,
  x = "t",
  y,
  linewidth = 1,
  var_linetype = NULL,
  manual_linetype = NULL,
  var_color = NULL,
  palette_color = NULL,
  manual_color = NULL,
  na.rm = TRUE,
  x_breaks = NULL,
  y_breaks = NULL,
  x_title = "",
  y_title = "",
  title = "",
  subtitle = "",
  caption = "",
  color_legend = "",
  linetype_legend = "",
  type_theme = "bw",
  var_facet = NULL,
  path_output = NULL,
  width = 15,
  height = 8,
  print = TRUE,
  return_output = TRUE
)
}
\arguments{
\item{baci}{Les données de la demande adressée. Il est possible de passer
un chemin d'accès vers un fichier csv ou parquet ou un dataframe R.}

\item{x}{La variable à mettre en abscisse. Par défaut, c'est le temps : \code{t}}

\item{y}{La variable à mettre en ordonnée. Il n'y a pas de paramètre par
défaut, la variable choisie étant dépendante du type de demande adressée
calculée.}

\item{linewidth}{L'épaisseur de la ligne du graphique. Par défaut, c'est 1.}

\item{var_linetype}{La variable qui va définir le type de ligne. Par défaut,
c'est \code{NULL}, les lignes seront toutes les mêmes.}

\item{manual_linetype}{Un vecteur indiquant à quelle entité quel type de ligne
doit correspondre.}

\item{var_color}{La variable à mettre en couleur. Par défaut, c'est \code{NULL}.}

\item{palette_color}{La palette de couleur à utiliser. Par défaut, c'est \code{NULL}.
Aucune palette de couleur ne sera utilisée/}

\item{manual_color}{Les couleurs à utiliser. Par défaut, c'est \code{NULL}. Aucune
couleur ne sera utilisée.}

\item{na.rm}{Indique si les valeurs manquantes doivent être retirées. Par
défaut, c'est \code{TRUE}.}

\item{x_breaks}{Les breaks à utiliser sur l'axe des x. Par défaut, c'est \code{NULL}.
dans ce cas, les breaks vont du min au max avec un écart de 2 entre chaque valeur.}

\item{y_breaks}{Les breaks à utiliser sur l'axe des y. Par défaut, c'est \code{NULL}.
Dans ce cas, les valeurs sont prises par défaut.}

\item{x_title}{Le titre de l'axe des x. Par défaut, c'est \code{""}.}

\item{y_title}{Le titre de l'axe des y. Par défaut, c'est \code{""}.}

\item{title}{Le titre du graphique. Par défaut, c'est \code{""}.}

\item{subtitle}{Le sous-titre du graphique. Par défaut, c'est \code{""}.}

\item{caption}{La caption du graphique. Par défaut, c'est \code{""}.}

\item{color_legend}{Le titre de la légende des couleurs. Par défaut, c'est \code{""}.}

\item{linetype_legend}{Le titre de la légende des types de ligne. Par défaut,
c'est \code{""}.}

\item{type_theme}{Le type de thème à utiliser. Par défaut, c'est \code{bw}. Les
autres options sont \code{classic} et \code{minimal}.}

\item{var_facet}{La variable à utiliser pour les facettes. Par défaut, c'est
\code{NULL}. Si utilisé, le paramètres \code{scales = "free_y"} est utilisé.}

\item{path_output}{Le chemin d'accès pour sauvegarder le graphique. Par défaut,
c'est \code{NULL}.}

\item{width}{La largeur du graphique. Par défaut, c'est 15.}

\item{height}{La hauteur du graphique. Par défaut, c'est 8.}

\item{print}{Indique si le graphique doit être affiché. Par défaut, c'est \code{TRUE}.}

\item{return_output}{Indique si le graphique doit être retourné. Par défaut,
c'est \code{TRUE}.}
}
\value{
Un graphique représentant l'évolution de la demande adressée.
}
\description{
Cette fonction permet de générer un graphique représentant l'évolution de la
demande adressée en fonction du temps. Selon les variables indiquées en input,
la demande adressée peut être en valeur, en base 100 ou par rapport à un
pays de référence.
}
\details{
La fonction permet de créer un graphique pour représenter l'évolution de la
demande adressée en fonction du temps. Elle utilise les fonctions de ggplot2
pour la création du graphique.

Il est important de noter, que cette fonction ne calcule pas la demande
adressée. Celle-ci doit être calculées précédemment (de préférence avec
la fonction \link{adressed_demand}). Il est possible de passer d'autres variables
que celles définies à partir des fonctions de ce package. Néanmoins cela reste
aux risques et périls de l'utilisateur si le rendu n'est pas cohérent.

Grâce au paramètre \code{return_output}, il est possible de récupérer le graphique
afin de pouvoir le modifier presque entièrement par la suite.
}
\examples{
# Pas d'exemples.
}
