% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_market_share.R
\name{graph_market_share}
\alias{graph_market_share}
\title{Graphique d'évolution des parts de marché}
\usage{
graph_market_share(
  baci,
  x = "t",
  y = "market_share",
  graph_type = "area",
  var_fill_color = NULL,
  palette_color = NULL,
  manual_color = NULL,
  percent = TRUE,
  na.rm = TRUE,
  x_breaks = NULL,
  y_breaks = NULL,
  x_title = "",
  y_title = "",
  title = "",
  subtitle = "",
  caption = "",
  color_legend = "",
  fill_legend = "",
  type_theme = "bw",
  var_facet = NULL,
  path_output = NULL,
  width = 15,
  height = 8,
  print = TRUE,
  return_output = TRUE
)
}
\arguments{
\item{baci}{Les données de BACI sous forme de chemin d'accès, de format
parquet ou de dataframe R.}

\item{x}{La variable correspondant à l'axe des abscisses. Par défaut et de
manière conseillé cette variable est "t.}

\item{y}{La variable correspondant à l'axe des ordonnées. Par défaut et de
manière conseillé cette variable est "market_share".}

\item{graph_type}{Le type de graphique à réaliser. Par défaut, le graphique
est de type "area". Les autres options sont "line" et "line_point". Pour
une représentation des parts de marché avec des lignes et potentiellement
des points sur les observations, il est conseillé de choisir area qui produit
un graphique en aire.}

\item{var_fill_color}{La variable catégorielle correspondant aux
différentes catégories à mettre en couleur.
Par défaut, cette variable est NULL.}

\item{palette_color}{La palette de couleur à utiliser pour le graphique.
Attention si le nombre de couleur dans la palette n'est pas suffisant par
rapport au nombre de catégories, une erreur sera retournée. Par défaut,
cette variable est NULL.}

\item{manual_color}{Les couleurs manuelles à utiliser pour le graphique.
Doit généralement être un vecteur contenant les couleurs utilisés et les
pays auxquels elles sont associées. Par défaut, cette variable est NULL.}

\item{percent}{Si TRUE, les valeurs de l'axe des ordonnées seront exprimées
en pourcentage. Par défaut, cette variable est TRUE. Attention, les valeurs
ne doit pas être exprimées de base de proportion, mais bien en pourcentage
(de 1 à 100).}

\item{na.rm}{Si TRUE, les valeurs manquantes seront retirées du graphique.}

\item{x_breaks}{Les valeurs à afficher sur l'axe des abscisses. Par défaut,
cette variable est NULL. Dans ce cas, les valeurs vont du minimum au maximum
de la variable x avec un écart de 2 entre chaque valeur.}

\item{y_breaks}{Les valeurs à afficher sur l'axe des ordonnées. Par défaut,
cette variable est NULL. Dans ce cas, les valeurs vont de 0 à 100 avec un
écart de 25 entre chaque valeur.}

\item{x_title}{Le titre de l'axe des abscisses. Par défaut, cette variable
est vide.}

\item{y_title}{Le titre de l'axe des ordonnées. Par défaut, cette variable
est vide.}

\item{title}{Le titre du graphique. Par défaut, cette variable est vide.}

\item{subtitle}{Le sous-titre du graphique. Par défaut, cette variable est
vide.}

\item{caption}{La légende du graphique. Par défaut, cette variable est vide.}

\item{color_legend}{Le titre de la légende des couleurs. Par défaut, cette
variable est vide. Inutilisé si graph_type est "area".}

\item{fill_legend}{Le titre de la légende des couleurs de remplissage. Par
défaut, cette variable est vide. Inutilisé si graph_type est "line" ou
"line_point".}

\item{type_theme}{Le thème général du graphique. Par défaut, cette variable
est "bw". Les valeurs possibles sont "bw", "classic" et "minimal".}

\item{var_facet}{La variable à utiliser pour les facettes. Par défaut, cette
variable est NULL. Si NULL alors l'entièreté des données sera représentée
sur un seul graphique. Sinon, les données seront séparées en fonction des
différentes catégories présentes dans la variable (généralement une variable
de secteur).}

\item{path_output}{Le chemin d'accès pour sauvegarder le graphique. Par
défaut, cette variable est NULL. Si NULL, le graphique ne sera pas sauvegardé.}

\item{width}{La largeur du graphique. Par défaut, cette variable est 15.}

\item{height}{La hauteur du graphique. Par défaut, cette variable est 8.}

\item{print}{Si TRUE, le graphique sera affiché. Par défaut, cette variable
est TRUE.}

\item{return_output}{Si TRUE, le graphique sera retourné permettant sa
modification par l'utilisateur. Par défaut, cettevariable est TRUE.}
}
\value{
Un graphique représentant l'évolution des parts de marché en fonction
du temps.
}
\description{
Fonction qui permet de créer un graphique pour représenter l'évolution des
parts de marché en fonction du temps. Le graphique peut être de type "area",
ou line avec ou sans point sur les observations.
}
\details{
La fonction permet de créer un graphique pour représenter l'évolution des
parts de marché en fonction du temps. Cete fonction utilise les fonctions
de ggplot2 pour la création du graphique.

Il est important de noter, que cette fonction ne calcule pas les parts de
marché. Celles-ci doivent être calculées précédemment (de préférence avec
la fonction \link{market_share}). Il est possible de passer d'autres variables
celles définies à partir des fonctions de ce package. Néanmoins cela reste
aux risques et périls de l'utilisateur si le rendu n'est pas cohérent.

Grâce au paramètre \code{return_output}, il est possible de récupérer le graphique
afin de pouvoir le modifier presque entièrement par la suite.
}
\examples{
# Pas d'exemple
}
