% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\name{market_share}
\alias{market_share}
\title{Calculer des parts de marché}
\usage{
market_share(
  baci,
  summarize_k = "k",
  summarize_v = "exporter",
  by = NULL,
  seuil = 0,
  years = NULL,
  codes = NULL,
  path_output = NULL,
  return_output = FALSE,
  return_pq = FALSE
)
}
\arguments{
\item{baci}{Peut être un  chemin d'accès vers le dossier contenant
les données de BACI au format parquet. Peut également être un dataframe ou
bien des données au format arrow (requête ou non) permettant ainsi de chaîner
les opérations entre elles. ce paramètre est obligatoire.}

\item{summarize_k}{Nom de la variable "produit" permettant à partir de
laquelle sera calculée la part de marché. Par c'est la avriable \code{k} (le code
produit HS6) qui sera utilisé. On peut l'utiliser également pour une
variable "chapitre" ou "secteur" par exemple.}

\item{summarize_v}{Nom de la variable ( en chaîne de caractère) sur laquelle
l'agrégation des flux va s'effectuer. Par défaut, c'est la variable
\code{exporteur} qui est prise en compte.}

\item{by}{Nom de la variable (en chaîne de caractère) sur laquelle les parts
de marché peuvent être calculées à un niveau plus fin. Par défaut, les parts
de marché sont calculées au niveau année-produit. Ce paramètre permet de
rajouter une dimmension, qui sera classiquement \code{importeur}.}

\item{seuil}{Valeur numérique entre 0 et 100 qui permet de filtrer les
données dans le dataframe en fonction de la part de marché. Seules les
observations ayant une part de marché supérieure ou égale à ce seuil seront
gardées. Par défaut, le seuil est de 0 : toutes les observations sont
gardées.}

\item{years}{Vecteur de numériques qui permet de filtrer les données en
fonction des années. Par défaut, toutes les années sont gardées.}

\item{codes}{Vecteur de chaînes de caractères qui permet de filtrer les
données en fonction des codes produits. Par défaut, tous les codes sont
gardés.}

\item{path_output}{Chemin d'accès vers le fichier où les données seront
sauvegardées. Par défaut, les données ne sont pas sauvegardées. Les formats
possibles sont .csv, .xlsx et .parquet.}

\item{return_output}{Booléen qui permet de retourner les données. Par défaut, les
données ne sont pas retournées.}

\item{return_pq}{Booléen pour indiquer si les données doivent être retournées
en format arrow si TRUE. Par défaut : FALSE.}
}
\value{
Un dataframe avec les parts de marché calculées.
}
\description{
Fonction qui permet de calculer les parts de marché des exportateurs pour
chaque produit / couple produit-importateur. Cette utilisation est celle qui
a été pensé à la création de la fonction. Mais elle peut être utilisée pour
calculer les parts de marché des importateurs pour chaque produit par exemple.
La définition de la part de marché (de qui, pour qui...) dépend des noms de
colonnes entrées en argument de la fonction.
}
\details{
La fonction va agréger les flux pour chaque couple année-produit-\code{summarize_v}
puis à partir de cette agrégation calculer les parts de marché. De base, ces
parts sont calculées pour chaque produit-année. Mais avec le paramètre \code{by},
il est possible de changer la définition de la part de marché. Par exemple,
si \code{by = "importer"}, les parts de marché seront calculées pour chaque
année-produit-importateur. Elle sera alors interprétée comme étant la part
de marché de l'exportateur sur le marché du produit pour cet importateur.

Les parts de marchés sont de base pensées et calculées avec les exportateurs
étant les pays définis par les variables \code{i} ou \code{exporteur}. Mais si la
variable d'intérêt est une variable de groupement régional par exemple,
il suffit de rentrer son nom pour que les parts de marché soient calculées.

Cependant, les parts de marché ne peuvent être calculées que sur un 'niveau'
après le produit : des parts de marché calculées au niveau
produit-région-importateur ne sont pas possibles. De même, les parts de
marché prennent toujours en compte la dimmension temps et produit.

Cette fonction utilise les fonctionnalité du package
\href{https://arrow.apache.org/docs/r/}{arrow} pour performer des calculs sans
avoir à charger BACI en mémoire. Cependant le calcul de la médiane
pondérée nécessite le passage de la base (uniquement la partie nécessaire)
en mémoire. Si la base est trop importante, les calculs peuvent prendre un
certain temps, voir entraîner un problème de mémoire de l'ordinateur. Si
cela arrive, il est conseillé de réduire le nombre d'années sur lesquelles
la fonction doit calculer les gammes et d'exécuter plusieurs fois la fonction
jusqu'à avoir toutes les années voulues.
}
\examples{
# Pas d'exemple.
}
