% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamme_ijkt_berthou_2011.R
\name{weighted_geomean}
\alias{weighted_geomean}
\title{Calcul une moyenne géométrique pondérée.}
\source{
rogiersbart with his RTOOLZ package : https://rdrr.io/github/rogiersbart/rtoolz/man/weighted.geomean.html
}
\usage{
weighted_geomean(x, w, ...)
}
\arguments{
\item{x}{Un vecteur de données numériques.}

\item{w}{Un vecteur de poids.}

\item{...}{Arguments supplémentaires à passer à la fonction \code{prod}.}
}
\value{
La moyenne géométrique pondérée.
}
\description{
Calcul une moyenne géométrique pondérée.
}
\examples{
# Pas d'exemple.
}
